### 2023-05-03 #5432 @editorialbot
[PRE REVIEW]: Scanbot: An STM Automation Bot
**Submitting author:** @ceds92 (<a href="http://orcid.org/0000-0003-3990-8852">Julian Ceddia</a>)
**Repository:** https://github.com/New-Horizons-SPM/scanbot
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.0.0
