### 2023-05-22 #5432 @ceds92
Hi @zhubonan,

Testing this software requires Spec's Nanonis V5 which is a licensed product. Further, it may be challenging for reviewers to find time in their schedules to pause their research and test this software on a live system (STM). However, this Scanbot can be tested on the STM simulator V5 software provided by Specs [here](https://www.specs-group.com/nanonis/products/mimea/). All functionality except for Scanbot's control over the STM tip via the motors + camera feed can be tested using this simulator.
For tracking the motion of the tip through the camera feed, I can upload a video recording of the STM tip moving, which can be used to replace the camera feed by running Scanbot in a 'demo' mode.

Some potential reviewers could be members of the following groups:
Michael Crommie (UBC)
Sarah Burke (UBC)
Doug Bonn (UBC)
Robert Wolkow (Alberta)
Peter Liljeroth (Aalto)
David Ecija (IMDEA)
Willi Auwärter (TUM)

Let me know if there's anything else I can help with!
