### 2023-08-07 3644e73
V3 Initial Commit
Added holoviz panel:
New GUI to interact with scanbot via panel.py.
Implemented so far:
 - survey
 - bias dependent imaging

Added:
panel.py: GUI... scanbot can now be run with ```panel serve panel.py --show```

Updated:
- pyproject.toml: new dependency for Panel
- scanbot.py: use the backend matplotlib
- scanbot_interface.py: added logic for run_mode = 'p' which triggers an update in panel.py
