+++ scanbot/panel.py
@@ -1,0 +1,220 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Aug  3 08:43:33 2023
+
+@author: jced0001
+"""
+
+import panel as pn
+from scanbot_interface import scanbot_interface
+import param
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib
+from collections import OrderedDict
+matplotlib.use('agg')
+
+from PIL import Image
+from pathlib import Path
+
+class ScanbotPanel(param.Parameterized):
+    template = 'fast'
+    sidebarColumn = []
+    function = {}
+    running = ""
+    prev_surveyForm  = {}
+    prev_biasdepForm = {}
+    tempFolder = "temp/"
+    
+    def __init__(self):
+        pn.extension(template=self.template)
+        self.interface = scanbot_interface(run_mode='p',panel=self)
+        self.initFunctions()
+        Path(self.tempFolder).mkdir(exist_ok=True)
+    
+    def initFunctions(self):
+        options = ['Configuration','Survey','Bias Dependent']
+        # Connection
+        self.functionWidget = pn.widgets.Select(name='Select', options=options)
+        interactive = pn.bind(self.selectFunction,self.functionWidget)
+        
+        self.sidebarColumn = pn.Column(self.functionWidget,interactive)
+        self.sidebarColumn.servable(target="sidebar")
+        
+        self.selectFunction(name=options[0])
+        
+        self.mainGridSpec = pn.GridSpec(sizing_mode='stretch_both',mode='override')
+        self.mainGridSpec.servable(target="main")
+        # self.mainGridSpec[0, :3] = pn.Spacer(styles=dict(background='#FF0000'))
+        
+    
+    def selectFunction(self,name):
+        while(len(self.sidebarColumn) > 2):
+            self.sidebarColumn.pop(-1)
+        
+        self.sidebarForm = []
+        if(name == 'Configuration'):
+            self.sidebarForm = self.getConnectionForm()
+        
+        if(name == 'Survey'):
+            self.sidebarForm = self.getSurveyForm()
+        
+        if(name == 'Bias Dependent'):
+            self.sidebarForm = self.getBiasDepForm()
+        
+        if(len(self.sidebarForm.keys())):
+            for f in self.sidebarForm.values(): self.sidebarColumn.append(f)
+        
+    def getConnectionForm(self):
+        fields = {}
+        fields['IP']            = self.interface.IP
+        fields['Ports']         = ','.join(np.array(self.interface.portList).astype(str))
+        fields['Upload Method'] = self.interface.uploadMethod
+        fields['Path']          = self.interface.path
+        fields['Whitelist']     = ','.join(np.array(self.interface.whitelist))
+        if(not fields['Whitelist']): fields['Whitelist'] = "Open"
+        fields['Crash Safety']  = "\n" + self.interface.getCrashSafety([])
+        
+        return {'fields': pn.panel(fields)}
+    
+    def getBiasDepForm(self):
+        if(self.prev_biasdepForm):
+            form = self.prev_biasdepForm.copy()
+            return form
+        
+        form = {}
+        options = list(np.arange(20)+1)
+        form['-n']     = pn.widgets.Select(name='Number of images', options=options)
+        form['-bi']    = pn.widgets.TextInput(name='Initial bias', value="-1")
+        form['-bf']    = pn.widgets.TextInput(name='Final bias', value="1")
+        form['-px']    = pn.widgets.TextInput(name='Pixels in data image', value="256")
+        form['-lx']    = pn.widgets.TextInput(name='Lines in data image (0 = same as pixels)', value="0")
[…]
