### 2023-08-07 b6fdf41
More features implemented
Updated:
 - panel.py: Added STM Control (move area + tip shaping)
 - scanbot_interface.py: change z3 to z2 in tipShapeProps
