+++ scanbot/panel.py
@@ -24,6 +24,7 @@ class ScanbotPanel(param.Parameterized):
     running = ""
     prev_surveyForm  = {}
     prev_biasdepForm = {}
+    prev_STMControlForm = {}
     tempFolder = "temp/"
     
     def __init__(self):
@@ -33,7 +34,7 @@ class ScanbotPanel(param.Parameterized):
         Path(self.tempFolder).mkdir(exist_ok=True)
     
     def initFunctions(self):
-        options = ['Configuration','Survey','Bias Dependent']
+        options = ['Configuration','Survey','Bias Dependent', 'STM Control']
         # Connection
         self.functionWidget = pn.widgets.Select(name='Select', options=options)
         interactive = pn.bind(self.selectFunction,self.functionWidget)
@@ -62,21 +63,106 @@ class ScanbotPanel(param.Parameterized):
         if(name == 'Bias Dependent'):
             self.sidebarForm = self.getBiasDepForm()
         
-        if(len(self.sidebarForm.keys())):
-            for f in self.sidebarForm.values(): self.sidebarColumn.append(f)
+        if(name == 'STM Control'):
+            self.sidebarForm = self.getSTMControlForm()
+            
+        for form in self.sidebarForm:
+            if(len(form.keys())):
+                for f in form.values(): self.sidebarColumn.append(f)
+    
+    def getSTMControlForm(self):
+        if(self.prev_STMControlForm): return self.prev_STMControlForm
+        
+        form1 = {}
+        form1['-up']    = pn.widgets.TextInput(name='Steps to take in Z+ before moving in X/Y', value="50")
+        form1['-steps'] = pn.widgets.TextInput(name='Steps to move across', value="20")
+        form1['-dir']   = pn.widgets.Select(name='Direction', options=["X+","X-","Y+","Y-"],value="X+")
+        form1['-upV']   = pn.widgets.TextInput(name='Piezo amplitude during Z+ steps (V)', value="180")
+        form1['-upF']   = pn.widgets.TextInput(name='Piezo frequency during Z+ steps (Hz)', value="2000")
+        form1['-dirV']  = pn.widgets.TextInput(name='Piezo amplitude during X/Y steps (V)', value="130")
+        form1['-dirF']  = pn.widgets.TextInput(name='Piezo frequency during X/Y steps (Hz)', value="2000")
+        form1['-zon']   = pn.widgets.Select(name='Turn z-controller on after move', options={"Turn on": 1,"Leave off": 0},value=0)
+        
+        buttonMove = pn.widgets.Button(name='Start Move', button_type='primary')
+        buttonMove.on_click(self.moveArea)
+        
+        form1['button1'] = pn.Row(buttonMove)
+        
+        form2 = {}
+        form2['-sod']   = pn.widgets.TextInput(name='Switch off delay (s)', value="0.1")
+        form2['-cb']    = pn.widgets.Select(name='Change bias?', options={"Yes":1,"No":0},value=0)
+        form2['-b1']    = pn.widgets.TextInput(name='B1: Bias to change to if yes (V)', value="0.4")
+        form2['-z1']    = pn.widgets.TextInput(name='Z1: First tip lift (m)', value="-2e-9")
+        form2['-t1']    = pn.widgets.TextInput(name='T1: Time to ramp Z1 (s)', value="0.1")
+        form2['-b2']    = pn.widgets.TextInput(name='B2: Bias applied just after the first Z ramping', value="-0.4")
+        form2['-t2']    = pn.widgets.TextInput(name='T2: Time to wait before second tip lift (s)', value="0.1")
[…]
