+++ README.md
@@ -1,3 +1,6 @@
+# V3 Update!
+Thanks to [holoviz Panel](https://panel.holoviz.org/), Scanbot now has a GUI. Clone the V3 branch, follow the normal installation, then run scanbot using ```panel serve panel.py --show```.
+
 # Scanbot       [![DOI](https://zenodo.org/badge/487719232.svg)](https://zenodo.org/badge/latestdoi/487719232)
 
 Scanbot is a collection of several automated STM and nc-AFM data acquisition commands compatible with Nanonis V5 SPM control software.
