### 2023-10-24 #5432 @zhubonan
Hi  @emilikv, @ziatdinovmax would you be interested to review this JOSS submission? It is about automating scanning tunnelling microscope and help for things like tip preparation.

The manuscript can be viewed here 👉 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.05432/joss.05432/10.21105.joss.05432.pdf)
Software repository 👉 https://github.com/New-Horizons-SPM/scanbot
The JOSS reviewer policy and reviewer guideline for your reference 👉 [Reviewer Guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks!
