### 2023-10-31 265b770
Bug fixes and more command additions
Fixed survey: now incorporates surveys at multiple locations (survey2 functionality)
Added auto stm functionality
