+++ scanbot/panel.py
@@ -18,23 +18,25 @@ from PIL import Image
 from pathlib import Path
 
 class ScanbotPanel(param.Parameterized):
-    template = 'fast'
     sidebarColumn = []
     function = {}
     running = ""
     prev_surveyForm  = {}
     prev_biasdepForm = {}
     prev_STMControlForm = {}
+    prev_AutomationForm = {}
     tempFolder = "temp/"
     
     def __init__(self):
-        pn.extension(template=self.template)
+        pn.extension(template="fast")
+        pn.config.template.title = "Scanbot"
         self.interface = scanbot_interface(run_mode='p',panel=self)
         self.initFunctions()
         Path(self.tempFolder).mkdir(exist_ok=True)
-    
+        
+        
     def initFunctions(self):
-        options = ['Configuration','Survey','Bias Dependent', 'STM Control']
+        options = ['Configuration','Survey','Bias Dependent', 'STM Control','Automation']
         # Connection
         self.functionWidget = pn.widgets.Select(name='Select', options=options)
         interactive = pn.bind(self.selectFunction,self.functionWidget)
@@ -46,7 +48,6 @@ class ScanbotPanel(param.Parameterized):
         
         self.mainGridSpec = pn.GridSpec(sizing_mode='stretch_both',mode='override')
         self.mainGridSpec.servable(target="main")
-        # self.mainGridSpec[0, :3] = pn.Spacer(styles=dict(background='#FF0000'))
         
     
     def selectFunction(self,name):
@@ -65,11 +66,82 @@ class ScanbotPanel(param.Parameterized):
         
         if(name == 'STM Control'):
             self.sidebarForm = self.getSTMControlForm()
+        
+        if(name == 'Automation'):
+            self.sidebarForm = self.getAutomationForm()
             
         for form in self.sidebarForm:
             if(len(form.keys())):
                 for f in form.values(): self.sidebarColumn.append(f)
     
+    def getAutomationForm(self):
+        if(self.prev_AutomationForm): return self.prev_AutomationForm
+        
+        form1 = {}
+        form1['-cameraPort'] = pn.widgets.Select(name='Camera port', options={"0":0,"1":1,"2":2,"3":3,"4":4},value=0)
+        form1['-demo']       = pn.widgets.Select(name='Demo/Live mode', options={"Live":0,"Demo":1},value=1)
+        
+        buttonInit = pn.widgets.Button(name='Initialise Tip Position', button_type='primary')
+        buttonInit.on_click(self.autoInit)
+        
+        form1['button1'] = pn.Row(buttonInit)
+        
+        form2 = {}
+        form2['-light']      = pn.widgets.Select(name='Control light with hk_light.py?', options={"No":0,"Yes":1},value=0)
+        form2['-cameraPort'] = pn.widgets.Select(name='Camera port', options={"0":0,"1":1,"2":2,"3":3,"4":4},value=0)
+        form2['-demo']       = pn.widgets.Select(name='Demo/Live mode', options={"Live":0,"Demo":1},value=1)
+        form2['-zStep']      = pn.widgets.TextInput(name='Steps at a time in Z+ direction', value="250")
+        form2['-zV']         = pn.widgets.TextInput(name='Piezo voltage when moving tip in Z+ (V)', value="150")
+        form2['-zF']         = pn.widgets.TextInput(name='Piezo frequency when moving tip in Z+ (Hz)', value="2000")
+        form2['-xStep']      = pn.widgets.TextInput(name='Steps at a time in X+/- direction', value="100")
+        form2['-xV']         = pn.widgets.TextInput(name='Piezo voltage when moving tip in X+/- (V)', value="80")
[…]
