### 2023-11-01 #5432 @zhubonan
Hi @jagar2 would you be interested to review this JOSS submission? It is about automating scanning tunnelling microscope and help for things like tip preparation.

The manuscript can be viewed here 👉 View article proof on GitHub
Software repository 👉 https://github.com/New-Horizons-SPM/scanbot
The JOSS reviewer policy and reviewer guideline for your reference 👉 Reviewer Guideline

Thanks!
