### 2023-11-01 #5432 @jagar2
I would be, but I am not going to be able to get to it for the next 3 weeks. Is that okay?

-------------------
Joshua C. Agar
Assistant Professor, Department of Mechanical Engineering and Mechanics
Drexel University
Website: www.m3-learning.com<http://www.m3-learning.com/>
3141 Chestnut Street
Philadelphia, PA 19104
Phone: 203-919-2230
Email: ***@***.******@***.***>
Pronouns: he, him, his

[Text, logo  Description automatically generated]

From: Bonan Zhu ***@***.***>
Sent: Tuesday, October 31, 2023 10:29 PM
Cc: Josh Agar ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Scanbot: An STM Automation Bot (Issue #5432)

External.

Hi @jagar2<https://github.com/jagar2> would you be interested to review this JOSS submission? It is about automating scanning tunnelling microscope and help for things like tip preparation.

The manuscript can be viewed here 👉 View article proof on GitHub
Software repository 👉 https://github.com/New-Horizons-SPM/scanbot
The JOSS reviewer policy and reviewer guideline for your reference 👉 Reviewer Guideline

Thanks!
