### 2023-11-01 3f18b7c
Minor bug fixes and added channel selection
