+++ scanbot/panel.py
@@ -36,6 +36,8 @@ class ScanbotPanel(param.Parameterized):
         
         
     def initFunctions(self):
+        self.generalInit()
+    
         options = ['Configuration','Survey','Bias Dependent', 'STM Control','Automation']
         # Connection
         self.functionWidget = pn.widgets.Select(name='Select', options=options)
@@ -49,7 +51,32 @@ class ScanbotPanel(param.Parameterized):
         self.mainGridSpec = pn.GridSpec(sizing_mode='stretch_both',mode='override')
         self.mainGridSpec.servable(target="main")
         
-    
+    def generalInit(self):
+        channels = self.interface.plotChannel(user_args=[]).split('\n')
+        # print("channels:",'\n'.join(channels))
+        self.channelBuffer = {}
+        self.availableChannels = {"Select": -1}
+        flag = 0
+        for channel in channels:
+            if('Channels in scan buffer' in channel): flag = 0; continue
+            if('Available channels' in channel): flag = 1; continue
+            if('Buffer idx' in channel): continue
+        
+            c = [i.strip() for i in channel.split('|')]
+            if(len(c) < 3): continue
+            if(c[2] in self.channelBuffer.keys()): continue
+        
+            if(c[2] == 'Z (m)'):
+                self.interface.plotChannel(['-a=' + c[0]])
+                self.interface.plotChannel(['-c=' + c[0]])
+                self.focusChannel = c[2]
+                self.channelBuffer[c[2]] = c[0]
+                continue
+            
+            if(flag): self.availableChannels[c[2]] = c[0]
+            if(not flag): self.channelBuffer[c[2]] = c[0]
+            
+                
     def selectFunction(self,name):
         while(len(self.sidebarColumn) > 2):
             self.sidebarColumn.pop(-1)
@@ -218,6 +245,8 @@ class ScanbotPanel(param.Parameterized):
         form['Ports']         = pn.widgets.TextInput(name='Port list (space delimited)', value=' '.join(np.array(self.interface.portList).astype(str)))
         form['Upload Method'] = pn.widgets.Select(name='Upload method', options=self.interface.validUploadMethods,value=self.interface.uploadMethod)
         form['Path']          = pn.widgets.TextInput(name='Save path', value=self.interface.path)
+        form['Focus Channel'] = pn.widgets.Select(name='Focus channel', options=self.channelBuffer,value=self.channelBuffer[self.focusChannel])
+        form['Add Channel']   = pn.widgets.Select(name='Add channel to buffer', options=self.availableChannels)
         # form['Crash Safety']  = pn.widgets.TextInput(name='IP Address', value=self.interface.IP)self.interface.getCrashSafety([])
         
         submitButton = pn.widgets.Button(name='Update Configuration', button_type='primary')
@@ -226,14 +255,40 @@ class ScanbotPanel(param.Parameterized):
         form['buttons'] = pn.Row(submitButton)
         
         return [form]
-    
-    def updateConfig(self,event):
-        config = self.sidebarForm[0]
-        self.interface.setIP([config['IP'].value])
-        self.interface.setPortList(config['Ports'].value.split(' '))
-        self.interface.setUploadMethod(config['Upload Method'].value)
-        self.interface.setPath([config['Path'].value])
         
+    def updateConfig(self,event):
+        error = []
+        
+        config = self.sidebarForm[0]
+        message = self.interface.setIP([config['IP'].value])
+        if(message): error.append("Error updating IP Adress. " + message)
+        message = self.interface.setPortList(config['Ports'].value.split(' '))
+        if(message): error.append("Error updating ports. " + message)
+        message = self.interface.setUploadMethod([config['Upload Method'].value])
+        if(message): error.append("Error updating upload type. " + message)
[…]
