### 2023-11-01 309491a
Update index.md
Instruction to run Scanbot using -gui
