### 2023-11-01 c97f14b
Automation improvements in GUI
Added auto tip shaping to GUI
minor bug fixes
