### 2023-11-01 9d04c0d
Added authentication to browser mode.
