### 2023-11-01 cdc10e5
Added documentation for auth
