+++ scanbot/scanbot.py
@@ -66,6 +66,15 @@ class scanbot():
 # Data Acquisition
 ###############################################################################
     def plot(self,channel=-1):
+        """
+        Return a plot of the currect scan frame
+
+        Parameters
+        ----------
+        channel : Channel in the buffer to plot. -1 means plot the channel 
+                  currently in focus.
+
+        """
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error): return connection_error                           # Return error message if there was a problem connecting        
         
@@ -90,6 +99,37 @@ class scanbot():
         self.disconnect(NTCP)                                                   # Close the TCP connection
             
     def survey(self,bias,n,startAt,suffix,xy,dx,px,sleepTime,stitch,survey_hk,classifier_hk,autotip,ox=0,oy=0,message="",enhance=False,reverse=False,iamauto=False):
+        """
+        This function carries out an autonomous survey within the scannable 
+        area.
+
+        Parameters
+        ----------
+        bias : Image bias (V)
+        n : Size of the grid of images (nxn)
+        startAt : Image index to start the survey at. (0 = first image)
+        suffix : Suffix appended to the .sxm filename for images acquired
+        xy : Width and height of the scan frame (m)
+        dx : Distance between scans in the grid (m)
+        px : Number of pixels and lines in each scan
+        sleepTime : Restart each scan after this amount of time to wait for
+                    drift to settle (s)
+        stitch : Save a stitched version of the survey after completing.
+        survey_hk : Call to hk_survey.py after each image completed. See 
+                    documentation.
+        classifier_hk : Call to hk_classifier after each scan. Used when auto
+                        shaping. See documentation for details.
+        autotip : Analayse scan frames and automatically navigate the tip to 
+                  a clean reference metal for tip shaping.
+        ox : Absolute x-centre coordinate for the survey grid (m)
+        oy : Absolute y-centre coordinate for the survey grid (m)
+        message : (Zulip use only)
+        enhance : (Not used in this version)
+        reverse : Reverse the direction of the survey (start from the end)
+        iamauto : Flag when this function was called autonomously
+
+        """
+        
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error):
             global_.running.clear()                                             # Free up the running flag
@@ -233,6 +273,11 @@ class scanbot():
         
     def survey2(self,bias,n,startAt,suffix,xy,dx,px,sleepTime,stitch,survey_hk,classifier_hk,autotip, # Survey params
                      nx,ny,xStep,yStep,zStep,xyV,zV,xyF,zF,message=""):          # Move area params
+        """
+        This function acquires many surveys with a call to move_area between
+        each one.
+        
+        """
         
         self.survey2Params = [bias,n,startAt,suffix,xy,dx,px,sleepTime,stitch,survey_hk,classifier_hk,autotip, # Survey2 params
                               nx,ny,xStep,yStep,zStep,xyV,zV,xyF,zF]
@@ -297,6 +342,37 @@ class scanbot():
             self.interface.moveTipToClean(user_args=user_args)
     
     def biasDep(self,nb,dcbias,tdc,dcSpeedRatio,pxdc,lxdc,bi,bf,px,lx,tlf,speedRatio,suffix,message=""):
+        """
+        This function initiates a set of bias dependent scans with the option 
+        of performing drift correction.
+        
+        When drift correction is turned on, every second scan is taken at a 
+        constant bias and used to determine a new scan frame position to 
+        accommodate for drift.
+        
+        Parameters
+        ----------
+        nb : Number of biases to scan area
+        dcbias : Bias of the drift correction scan frame. Set to 0 to turn off
[…]
