+++ scanbot/scanbot.py
@@ -25,6 +25,8 @@ import ntpath
 import numpy as np
 import nanonispyfit as napfit
 import matplotlib.pyplot as plt
+import matplotlib
+matplotlib.use('agg')
 import cv2
 
 import math
@@ -64,6 +66,15 @@ class scanbot():
 # Data Acquisition
 ###############################################################################
     def plot(self,channel=-1):
+        """
+        Return a plot of the currect scan frame
+
+        Parameters
+        ----------
+        channel : Channel in the buffer to plot. -1 means plot the channel 
+                  currently in focus.
+
+        """
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error): return connection_error                           # Return error message if there was a problem connecting        
         
@@ -88,6 +99,37 @@ class scanbot():
         self.disconnect(NTCP)                                                   # Close the TCP connection
             
     def survey(self,bias,n,startAt,suffix,xy,dx,px,sleepTime,stitch,survey_hk,classifier_hk,autotip,ox=0,oy=0,message="",enhance=False,reverse=False,iamauto=False):
+        """
+        This function carries out an autonomous survey within the scannable 
+        area.
+
+        Parameters
+        ----------
+        bias : Image bias (V)
+        n : Size of the grid of images (nxn)
+        startAt : Image index to start the survey at. (0 = first image)
+        suffix : Suffix appended to the .sxm filename for images acquired
+        xy : Width and height of the scan frame (m)
+        dx : Distance between scans in the grid (m)
+        px : Number of pixels and lines in each scan
+        sleepTime : Restart each scan after this amount of time to wait for
+                    drift to settle (s)
+        stitch : Save a stitched version of the survey after completing.
+        survey_hk : Call to hk_survey.py after each image completed. See 
+                    documentation.
+        classifier_hk : Call to hk_classifier after each scan. Used when auto
+                        shaping. See documentation for details.
+        autotip : Analayse scan frames and automatically navigate the tip to 
+                  a clean reference metal for tip shaping.
+        ox : Absolute x-centre coordinate for the survey grid (m)
+        oy : Absolute y-centre coordinate for the survey grid (m)
+        message : (Zulip use only)
+        enhance : (Not used in this version)
+        reverse : Reverse the direction of the survey (start from the end)
+        iamauto : Flag when this function was called autonomously
+
+        """
+        
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error):
             global_.running.clear()                                             # Free up the running flag
@@ -186,6 +228,7 @@ class scanbot():
                     global_.running.clear()
                     callAutoTipShape = True
             
+            _,scanData,_ = scan.FrameDataGrab(self.channel, 1)                  # Grab the data related to our focused channel
             pngFilename,scanDataPlaneFit = self.makePNG(scanData, filePath,returnData=True,dpi=150) # Generate a png from the scan data
             self.interface.sendPNG(pngFilename,notify=True,message=message)     # Send a png over zulip
             
@@ -211,7 +254,7 @@ class scanbot():
             if(self.checkEventFlags()): break                                   # Check event flags
         
         if(stitch == 1 and not np.isnan(stitchedSurvey).all()):
-            stitchFilepath = self.makePNG(stitchedSurvey,pngFilename = suffix + '.png',dpi=150*n, fit=False)
+            stitchFilepath = self.makePNG(stitchedSurvey,pngFilename = suffix + '_stitch.png',dpi=150*n, fit=False)
             self.interface.sendPNG(stitchFilepath,notify=False,message=message) # Send a png over zulip
         
[…]
