### 2023-11-03 a9c1584
Improvements to demo mode and gui
