### 2023-11-03 bb3500e
Improvements to demo mode
