+++ scanbot/scanbot.py
@@ -35,6 +35,18 @@ import global_
 
 import utilities
 
+try: import hk_classifier
+except: pass
+
+try: import hk_tipShape
+except: pass
+
+try: from hk_light import turn_on
+except: pass
+
+try: import pickle
+except: pass
+
 class scanbot():
     channel      = 14                                                           # Default plot channel. Change this using plot_channel command
     
@@ -214,8 +226,7 @@ class scanbot():
                 classification = utilities.classify(scanData,filePath,classificationHistory) # Obtain image classification
                 if(classifier_hk):
                     try:
-                        from hk_classifier import run
-                        classification = run(scanData,filePath,classificationHistory) # Overwrite classification with the one from the hook
+                        classification = hk_classifier.run(scanData,filePath,classificationHistory) # Overwrite classification with the one from the hook
                     except Exception as e:
                         self.interface.sendReply("Warning: Call to survey hook hk_classifier.py failed:")
                         self.interface.sendReply(str(e))
@@ -1114,7 +1125,6 @@ class scanbot():
         """
         if(lightOnOff):                                                         # If we want to turn the light on
             try:
-                from hk_light import turn_on
                 turn_on()                                                       # Call the hook to do so. Hook should return null if successful, otherwise it should throw an Exception
             except Exception as e:
                 self.interface.sendReply("Error calling hook hk_light.py to turn on light")
@@ -1441,7 +1451,7 @@ class scanbot():
         global_.running.clear()                                                 # Free up the running flag
         return
     
-    def moveTipToTarget(self,lightOnOff, cameraPort, xStep, zStep, xV, zV, xF, zF, approach, demo, tipshape, retrn, run, target):
+    def moveTipToTarget(self,lightOnOff, cameraPort, xStep, zStep, xV, zV, xF, zF, approach, demo, tipshape, tipShape_hk, retrn, run, target, autoTipParams=[]):
         if(not self.autoInitSet):
             self.interface.sendReply("Error, run the auto_init command to initialise tip, sample, and clean metal locations")
             global_.running.clear()                                             # Free up the running flag
@@ -1461,7 +1471,7 @@ class scanbot():
         
         message = ""
         if(target == "clean" and tipshape == 1):
-            message = self.autoTipShape(n=50, wh=10e-9, symTarget=0.7, sizeTarget=3, zQA=-85e-11, ztip=-2.5e-9, sleepTime=1, iamauto=True, demo=demo)
+            message = self.autoTipShape(n=20, wh=10e-9, symTarget=0.7, sizeTarget=3, zQA=-85e-11, ztip=-2.5e-9, sleepTime=1, iamauto=True, demo=demo, tipShape_hk=tipShape_hk)
             if(not message): message = ""
         
         if(not "Tip shaping successful" in message):
@@ -1491,7 +1501,7 @@ class scanbot():
             self.survey2(user_args=[],_help=False,surveyParams=self.survey2Params)
             return
         
-    def autoTipShape(self,n,wh,symTarget,sizeTarget,zQA,ztip,rng=1,sleepTime=1,tipShape_hk="",message="",iamauto=False,demo=False):
+    def autoTipShape(self,n,wh,symTarget,sizeTarget,zQA,ztip,rng=1,sleepTime=1,demo=False,tipShape_hk="",message="",iamauto=False):
         """
         This function initiates autonomous tip shaping. The process is as 
         follows:
@@ -1589,6 +1599,8 @@ class scanbot():
         if(rng == 1): rng = np.random.default_rng()
         else: rng = 0
         
+        hk_tipShapeHistory = []                                                 # Variable used to pass in and out of hk_tipShape
+        
         tipShapeProps[10]   = 1                                                 # Make sure feedback on after tip shape
         tipShapeProps[3]   = ztip                                               # Amount to dip the tip into the surface
[…]
