### 2023-11-04 3b88d50
Added featured + update hk_tipShape docs
