### 2023-11-07 #5432 @zhubonan
> Hi all,
>
> @zhubonan thanks for sticking with this and finding reviewers, it is much appreciated! @jagar2 @ziatdinovmax, thanks for agreeing to review!!
>
> Just a heads up... I have been doing some work on Scanbot over the few months since it was submitted. It has a GUI now which should make things easier when testing - this is available on the V3 branch. All core functionality is basically the same.
>
> Also, given that STM time is generally precious, I've created ways to test Scanbot's functionality, without needing an STM, through introducing a demo mode (along with nanonis V5 simulator). For example, when tracking the motion of the STM tip in demo mode, the camera feed will be replaced by a recording of the tip moving. The other main example is when automated tip shaping is initiated, STM images of previously acquired tip imprints will be loaded from a pickled file instead of using images from the simulator.
>
> The files for testing can be downloaded from [this link](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/Dev%2FDev.zip?alt=media&token=7cef378d-d7da-4851-8849-eaa73212d27f). Just save and extract it in Scanbot's root directory.
>
> If you have any questions, let me know!
>
> Thanks!!

@ceds92 Thanks for the update. I think we can start the review.
