### 2023-11-07 #6028 @editorialbot
[REVIEW]: Scanbot: An STM Automation Bot
**Submitting author:** @ceds92 (<a href="http://orcid.org/0000-0003-3990-8852">Julian Ceddia</a>)
**Repository:** https://github.com/New-Horizons-SPM/scanbot
**Branch with paper.md** (empty if default branch): joss
**Version:** 4.5.1
