### 2023-11-11 f75b62e
Bug fix in example ChatGPT code
