### 2023-11-11 b512778
Update featured.md
add openai compatible version to code comments
