### 2023-11-13 7b3a3bc
Update docs with GUI page
