+++ docs/gui.md
@@ -1,0 +1,156 @@
+# Scanbot GUI
+
+Thanks to [holoviz Panel](https://panel.holoviz.org/), Scanbot now has a GUI that runs in a browser.
+This is still under active development, so not all of Scanbot's functionality is accessible from the GUI yet. 
+Currently, the following functions have been implemented in the Scanbot GUI:
+
+* Survey
+* Bias Dependent Imaging
+* STM Control
+    - Move area
+    - Tip Shape
+    - Auto Tip Shape
+* End-To-End Automation
+
+## Overview
+
+Scanbot can be started in GUI mode by running the command ```python scanbot_interface.py -gui``` from ~/scanbot/scanbot
+The Scanbot GUI is split into two panels: the side panel and the plot area. The side panel is where functions and parameters can be selected,
+and the plot area is where messages and images are displayed.
+
+![survey_window](guiim/surveywindow.png)
+
+
+## Survey
+
+When the Survey function is selected, the following parameters can be configured:
+
+| Parameter                                            | Description                                                                                                                                                                                                                                                                                                                             |
+|------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| Survey fine grid size (NxN)                          | Surveys are grids of images. This parameter defines the grid size. For example, if N=3, the survey grid will be 9 images in a 3x3 grid                                                                                                                                                                                                  |
+| Survey course grid size X/Y                          | The course grid is a macroscopic grid of points on the sample where surveys will be carried out.<br>At the completion of each survey, the course piezos will move the tip to the next location in the course grid.<br>The course grid size is XxY.<br>For example, if X=5 and Y=2, a total of 10 fine-grid surveys will be carried out. |
+| Scan size (m)                                        | Length and widthof each image in the survey in meters                                                                                                                                                                                                                                                                                   |
+| Scan spacing (m)                                     | Spacing between images in a survey in meters                                                                                                                                                                                                                                                                                            |
+| Number of pixels                                     | Number of pixels in each image in the survey                                                                                                                                                                                                                                                                                            |
+| Scan bias (V)                                        | Bias the images in the survey will be acquired at                                                                                                                                                                                                                                                                                       |
+| Filename suffix                                      | String that will be appended to the end of the .sxm filenames for all images acquired in the survey                                                                                                                                                                                                                                     |
+| Drift compensation (s)                               | Number of seconds after which to restart each scan. This is to accommodate for drift induced by the scan frame moving.                                                                                                                                                                                                                  |
+| Number of motor steps (Z+)                           | Number of motor steps the tip will retract before moving to the next survey location in the course grid                                                                                                                                                                                                                                 |
+| Piezo voltage during Z+ motor steps (V)              | Voltage applied to the Z piezos when retracting the tip before moving to the next course grid location                                                                                                                                                                                                                                  |
+| Piezo frequency during Z+ motor steps (Hz)           | Frequency applied to the Z piezos when retracting the tip before moving to the next course grid location                                                                                                                                                                                                                                |
+| Number of motor steps (X/Y)                          | Number of piezo motor steps (after retracting in Z+) between course grid locations                                                                                                                                                                                                                                                      |
[…]
