### 2023-11-13 ef1181f
Update autotipshape.png
