+++ mkdocs.yml
@@ -6,6 +6,7 @@ nav:
     - Home: 'index.md'
     - Featured: 'featured.md'
     - Configuration: 'configuration.md'
+    - GUI: 'gui.md'
     - Commands: 'commands.md'
     - Automation: 'automation.md'
     - Hooks: 'hooks.md'
