+++ docs/contact.md
@@ -1,5 +1,6 @@
 # Contact
-
-Julian Ceddia: julian.ceddia@monash.edu
+<br>
+email: julian.ceddia@monash.edu
+website: [julianceddia.com](https://julianceddia.com)
 
 Any feedback would be much appreciated!
