+++ docs/about.md
@@ -1,19 +1,6 @@
-## About
-<br>
-Scanning Tunnelling Microscopes (STM) are powerful instruments, capable of characterising the topography and electronic properties of surfaces with atomic scale resolution.
-They accomplish this by scanning an atomically sharp probe at a small distance (< 1 nm) above the surface of a sample, while monitoring an electric current that tunnels across the gap.
-This tunnelling current depends on the nature of the sample directly below the tip, thus atomic-scale images of the surface can be assembled by plotting the tunnelling current as a function
-of the probe’s position in x and y. In this way, surfaces can be characterised from the microscopic scale (on the order of micrometres) down to the sub-nanometre scale.
-<br>
-<br>
-Conditioning (e.g. sharpening) of the probe, along with finding regions of interest on a sample are two time-consuming (and often tedious) tasks common in STM experiments.
-The quality of STM images depends greatly on the exact geometry and composition at the apex of the scanning probe. Blunt tips result in blurry images while contaminated tips can
-interact with the sample during a scan to produce noisy images. To improve image quality, the probe can be conditioned via a process called ‘tip shaping’, which involves plunging the
-tip of the probe into a metallic substrate in order to refine it. This is a somewhat random process that is repeated until a tip capable of acquiring images of sufficient quality is achieved.
-Maintaining a 'good tip' is not guaranteed. For instance, scanning over debris or excessively rough areas can alter the tip, resulting in the need to tip shape once again. This frustrating (and often infuriating)
-task is currently performed by a trained human operator who has a 'feel' for the instrument. Automating these processes has the potential to greatly increase the efficiency and output of STMs.
-<br>
-<br>
-This project aims to automate STM probe conditioning and sample surveying as a first step towards full automation in STM experiments. It was made possible by [FLEET](https://www.fleet.org.au/)
+# About
+This project automates STM probe conditioning and sample surveying as a first step towards full automation in STM experiments. It was made possible by [FLEET](https://www.fleet.org.au/)
 through their [Research Translation Program](https://www.fleet.org.au/translation/#:~:text=A%20new%20FLEET%20program%20provides,translation%20skills%20in%20Centre%20membership.).
-![FLEETLogo](fleet-logo.png)
+<br>
+## Poster
+![scanbotposter](scanbotposter.png)
