### 2023-12-12 c823195
Update to account for changes in hk_tipShape.py
