### 2024-01-15 1b75e2c
Update utilities.py
Fix drift correction (make it more robust)
