### 2024-01-15 6e1b0c2
Update utilities.py
remove debug from utilities
