### 2024-01-25 9e509d9
React implementation of Scanbot
In this version:
Completed:
- Surveying (including autotip)
- bias dependent imaging
- Configuration

ToDo:
Implement quick move and automatic tip shaping in /automation
