+++ scanbot/src/BiasDep.js
@@ -1,0 +1,269 @@
+import React, { useEffect, useState, useRef } from 'react';
+import { Sidebar, usePersistedState, GoBack } from './Components';
+import { positiveInt, number, positiveNumber } from './Validations';
+import emptyFrameIcon from './img/frame.png';
+import './styles/Survey.css'
+
+function BiasDep() {
+    const initialAllFormData = {
+        0: {n: '3', bi: '-1', bf: '1', s: 'scanbot-biasdep' },
+        1: {px: '0', lx: '0', tlf: '0', tb: '1' },
+        2: {bdc: '0.5', pxdc: '128', lxdc: '0', tdc: '0', tbdc: '1'  }
+    }
+    const [allFormData, setAllFormData] = usePersistedState('bias-dep-allFormData', initialAllFormData)
+    const [running, setRunning] = useState(false)
+    const [lastImage,  setLastImage]  = useState({ src: emptyFrameIcon, alt: "blank image", width: 300, height: 300} );
+    const [biasDepGif, setBiasDepGif] = useState([{ src: emptyFrameIcon, alt: "blank image", width: 300, height: 300}] );
+    const [imageIndex, setImageIndex] = useState(0);
+    const [lastTimestamp, setLastTimestamp] = useState(0);
+    const timerIdRef = useRef(null);
+    
+    const handleInputChange = (formIndex, name, value, index) => {
+        var goahead = true
+
+        // Validations
+        const positiveNumbers = ['tlf','tb','pxdc','lxdc','tdc','tbdc']
+        const positiveInts = ['n', 'px', 'lx']
+        const numbers = ['bi','bf', 'bdc']
+        if(positiveInts.includes(name)) {
+            goahead = positiveInt(value)
+        }
+        if(numbers.includes(name)) {
+            goahead = number(value)
+        }
+        if(positiveNumbers.includes(name)) {
+            goahead = positiveNumber(value)
+        }
+
+        if(goahead) {
+            setAllFormData({
+                ...allFormData,
+                [formIndex]: {
+                    ...allFormData[formIndex],
+                    [name]: value
+                }
+            });
+        }
+    };
+    
+    const handleFormSubmit = () => {
+        if(running){
+            setRunning(false)
+            stop()
+        }else{
+            startBiasDep()
+        }
+    };
+    
+    const stop = () => {
+        fetch("/stop")
+    }
+
+    const startBiasDep = () => {
+        const userArgs = unpackArgs()
+
+        fetch('/run_biasdep', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({userArgs: userArgs}),
+        })
+        .then(response => response.json())
+        .then(data => {
+            if(data['status'] === "success"){
+                var image = lastImage
+                image['src'] = emptyFrameIcon
+                if(lastImage){
+                    URL.revokeObjectURL(lastImage);
+                    URL.revokeObjectURL(biasDepGif[0]);
+                }
+        
+                setLastImage(image)
+                setBiasDepGif([image])
+                setImageIndex(0)
+                setRunning(true)
+                console.log('Success:', data);
+            }else{
+                console.log('Fail', data)
+            }
+            
+        })
+        .catch((error) => {
+            console.error('Error:', error);
+        });
+    }
+    
+    const unpackArgs = () => {
+        var unpackedArgs = []
+        const formIdx = Object.keys(allFormData)
+        for (let idx = 0; idx < formIdx.length; idx++) {
+            const form = allFormData[formIdx[idx]]
+            const keys = Object.keys(form)
+            for (let index = 0; index < keys.length; index++) {
+                const key = '-' + keys[index]
[…]
