+++ server/scanbot_interface.py
@@ -13,6 +13,8 @@ import os
 import sys
 import subprocess
 from pathlib import Path
+import shutil
+import time
 
 import ipaddress
 
@@ -259,7 +261,7 @@ class scanbot_interface(object):
                     '-dx'   : ['-default', lambda x: float(x), "(float) Scan grid spacing (m). Default is -xy"],
                     '-px'   : ['-default', lambda x: int(x),   "(int) Number of pixels"],
                     '-st'   : ['10',       lambda x: float(x), "(float) Drift compensation time (s)"],
-                    '-stitch':['1',        lambda x: float(x), "(int) Return the stitched survey after completion. 1: Yes, else No"],
+                    '-stitch':['0',        lambda x: float(x), "(int) Return the stitched survey after completion. 1: Yes, else No"],
                     '-hk_survey': ['0',    lambda x: int(x),   "(int) Flag to call a custom python script after each image. Script must be ~/scanbot/scanbot/hk_survey.py. 0=Don't call, 1=Call"],
                     '-hk_classifier': ['0',lambda x: int(x),   "(int) Flag to call a custom classifier after each image. Script must be ~/scanbot/scanbot/hk_classifier.py. Only called if -autotip=1. 0=Don't call, 1=Call"],
                     '-autotip': ['0',      lambda x: int(x),   "(int) Automatic tip shaping. 0=off, 1=on"]}
@@ -285,7 +287,7 @@ class scanbot_interface(object):
                     '-dx'   : ['-default', lambda x: float(x), "(float) Scan grid spacing (m)"],
                     '-px'   : ['-default', lambda x: int(x),   "(int) Number of pixels"],
                     '-st'   : ['10',       lambda x: float(x), "(float) Drift compensation time (s)"],
-                    '-stitch':['1',        lambda x: float(x), "(int) Return the stitched survey after completion. 1: Yes, else No"],
+                    '-stitch':['0',        lambda x: float(x), "(int) Return the stitched survey after completion. 1: Yes, else No"],
                     '-hk_survey': ['0',    lambda x: int(x),   "(int) Flag to call a custom python script after each image. Script must be ~/scanbot/scanbot/hk_survey.py. 0=Don't call, 1=Call"],
                     '-hk_classifier': ['0',lambda x: int(x),   "(int) Flag to call a custom classifier after each image. Script must be ~/scanbot/scanbot/hk_classifier.py. Only called if -autotip=1. 0=Don't call, 1=Call"],
                     '-autotip': ['0',      lambda x: int(x),   "(int) Automatic tip shaping. 0=off, 1=on"],
@@ -309,6 +311,9 @@ class scanbot_interface(object):
             if(error): return error + "\nRun ```help survey2``` if you're unsure."
             
             args = self.unpackArgs(user_arg_dict)
+            if(self.run_mode == 'react'):   # React interface passes scan size and spacing in units of nm. convert to m
+                args[4] *= 1e-9
+                args[5] *= 1e-9
         
         func = lambda : self.scanbot.survey2(*args,message=self.bot_message.copy())
         return self.threadTask(func)
@@ -453,16 +458,21 @@ class scanbot_interface(object):
     def autoInit(self,user_args,_help=False):
         arg_dict = {'-light'      : ['0',   lambda x: int(x),   "(int) Flag to turn the light on/off before/after initialisation. This uses hk_light.turn_on() and hk_light.turn_off() functions. 0=Don't, 1=Do"],
                     '-cameraPort' : ['0',   lambda x: int(x),   "(int) cv2 camera port - usually 0 for desktops or 1 for laptops with an inbuilt camera."],
-                    '-demo'       : ['0',   lambda x: int(x),   "(int) Load in an mp4 recording of the tip moving instead of using live feed"]}
[…]
