+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "opencv-python",
     "zulip",
     "panel",
+    "flask",
 ]
 
 [tool.setuptools]
