### 2024-01-25 3c0b75d
Update README.md
New installation instructions for v4
