### 2024-01-26 c7a82fc
Change scp in config file to not need flag
