+++ docs/web-app.md
@@ -1,0 +1,2 @@
+## Coming Soon
+Documentation coming soon.
