+++ docs/commands.md
@@ -1,4 +1,5 @@
 # Commands
+This page documents the available commands and how to use them when running Scanbot via Zulip or in a terminal.
 For a list of available commands on the fly, run ```help```, or for help with a specific command, run ```help <command_name>```
 <br>
 ## Scanbot Configuration
