### 2024-01-26 0f882fd
Add panel.py for running in browser mode
From V3
