+++ server/panel.py
@@ -1,0 +1,570 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Aug  3 08:43:33 2023
+
+@author: jced0001
+"""
+
+import panel as pn
+from scanbot_interface import scanbot_interface
+import param
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib
+from collections import OrderedDict
+matplotlib.use('agg')
+
+from PIL import Image
+from pathlib import Path
+
+class ScanbotPanel(param.Parameterized):
+    sidebarColumn = []
+    function = {}
+    running = ""
+    prev_surveyForm  = {}
+    prev_biasdepForm = {}
+    prev_STMControlForm = {}
+    prev_AutomationForm = {}
+    tempFolder = "temp/"
+    
+    def __init__(self):
+        pn.extension(template="fast")
+        pn.config.template.title = "Scanbot"
+        self.interface = scanbot_interface(run_mode='p',panel=self)
+        self.initFunctions()
+        Path(self.tempFolder).mkdir(exist_ok=True)
+        
+        
+    def initFunctions(self):
+        self.generalInit()
+    
+        options = ['Configuration','Survey','Bias Dependent', 'STM Control','Automation']
+        # Connection
+        self.functionWidget = pn.widgets.Select(name='Select', options=options)
+        interactive = pn.bind(self.selectFunction,self.functionWidget)
+        
+        self.sidebarColumn = pn.Column(self.functionWidget,interactive)
+        self.sidebarColumn.servable(target="sidebar")
+        
+        self.selectFunction(name=options[0])
+        
+        self.mainGridSpec = pn.GridSpec(sizing_mode='stretch_both',mode='override')
+        self.mainGridSpec.servable(target="main")
+        
+    def generalInit(self):
+        channels = self.interface.plotChannel(user_args=[]).split('\n')
+        # print("channels:",'\n'.join(channels))
+        self.channelBuffer = {}
+        self.availableChannels = {"Select": -1}
+        flag = 0
+        for channel in channels:
+            if('Channels in scan buffer' in channel): flag = 0; continue
+            if('Available channels' in channel): flag = 1; continue
+            if('Buffer idx' in channel): continue
+        
+            c = [i.strip() for i in channel.split('|')]
+            if(len(c) < 3): continue
+            if(c[2] in self.channelBuffer.keys()): continue
+        
+            if(c[2] == 'Z (m)'):
+                self.interface.plotChannel(['-a=' + c[0]])
+                self.interface.plotChannel(['-c=' + c[0]])
+                self.focusChannel = c[2]
+                self.channelBuffer[c[2]] = c[0]
+                continue
+            
+            if(flag): self.availableChannels[c[2]] = c[0]
+            if(not flag): self.channelBuffer[c[2]] = c[0]
+            
+                
+    def selectFunction(self,name):
+        while(len(self.sidebarColumn) > 2):
+            self.sidebarColumn.pop(-1)
+        
+        self.sidebarForm = []
+        if(name == 'Configuration'):
+            self.sidebarForm = self.getConnectionForm()
+        
+        if(name == 'Survey'):
+            self.sidebarForm = self.getSurveyForm()
+        
+        if(name == 'Bias Dependent'):
+            self.sidebarForm = self.getBiasDepForm()
+        
+        if(name == 'STM Control'):
+            self.sidebarForm = self.getSTMControlForm()
+        
+        if(name == 'Automation'):
+            self.sidebarForm = self.getAutomationForm()
+            
+        for form in self.sidebarForm:
+            if(len(form.keys())):
[…]
