### 2024-01-26 2c0beb5
Make zulip live stream work with v4
