+++ docs/web-app.md
@@ -1,2 +1,112 @@
-## Coming Soon
-Documentation coming soon.
+## Scanbot Web App
+
+Scanbot has been implemented as a web application using [React](https://react.dev/).
+This application is still being developed, with the following features already fully implemented:
+
+* Survey
+* Bias Dependent Imaging
+* End-To-End Survey Automation
+
+![WebApp](./appim/top-level.png)
+
+## Installation
+
+To use the web application, you can follow the normal installation, then complete these additional steps:
+
+    1. Install node.js from [here](https://nodejs.org/en) or if you're using anaconda, run conda install conda-forge::nodejs
+    2. Navigate to ```~/scanbot/scanbot``` and ```run npm install```
+    3. Start the server: navigate to ```~/scanbot/server/``` and run ```python server.py```
+    4. Start the web app: navigate to ```~/scanbot/scanbot/``` and run ```npm start```
+
+## Configuration
+From the landing page, you can access Scanbot's configuration.
+It is reccommended that you review the default settings before using any of the tools in Scanbot.
+Upon accepting the configuration, your scanbot_config.ini file will be saved automatically and your settings will be remembered.
+These settings can be updated any time. <strong>For a detailed description about the configurable parameters, see 
+[configuration](../configuration).</strong>
+![Config](./appim/configuration.png)
+
+## Data Acquisition
+From the landing page, you can access the data acquisition tools, including tools to help automate the STM.
+![DataAcquisition](./appim/data-acquisition.png)
+
+## Survey
+![DataAcquisition](./appim/survey.png)
+
+When the Survey function is selected, the parameters in the table below can be configured.
+Note that answering ```Yes``` to ```Auto tip shaping?```, is only possible after following a short
+[initialisation procedure](/stm-automation).
+
+| Parameter                                            | Description                                                                                                                                                                                                                                                                                                                             |
+|------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| Survey fine grid size (NxN)                          | Surveys are grids of images. This parameter defines the grid size. For example, if N=3, the survey grid will be 9 images in a 3x3 grid                                                                                                                                                                                                  |
+| Survey course grid size X/Y                          | The course grid is a macroscopic grid of points on the sample where surveys will be carried out.<br>At the completion of each survey, the course piezos will move the tip to the next location in the course grid.<br>The course grid size is XxY.<br>For example, if X=5 and Y=2, a total of 10 fine-grid surveys will be carried out. |
+| Scan size (m)                                        | Length and widthof each image in the survey in meters                                                                                                                                                                                                                                                                                   |
+| Scan spacing (m)                                     | Spacing between images in a survey in meters                                                                                                                                                                                                                                                                                            |
+| Number of pixels                                     | Number of pixels in each image in the survey                                                                                                                                                                                                                                                                                            |
+| Scan bias (V)                                        | Bias the images in the survey will be acquired at                                                                                                                                                                                                                                                                                       |
+| Filename suffix                                      | String that will be appended to the end of the .sxm filenames for all images acquired in the survey                                                                                                                                                                                                                                     |
+| Drift compensation (s)                               | Number of seconds after which to restart each scan. This is to accommodate for drift induced by the scan frame moving.                                                                                                                                                                                                                  |
+| Number of motor steps (Z+)                           | Number of motor steps the tip will retract before moving to the next survey location in the course grid                                                                                                                                                                                                                                 |
[…]
