### 2024-01-29 #6028 @ziatdinovmax
@zhubonan I'm working through the checklist.

@ceds92, can you please add unit tests and at least brief function documentation?
