### 2024-01-30 e751f15
Update pyproject.toml
open-cv latest version broke auto_init
