### 2024-02-09 #6028 @ceds92
Hi @ziatdinovmax, just wondering if the document I uploaded contains everything you asked for or if there's any additional information you wanted?
