### 2024-02-11 e889661
Docs updated for web-app and browser
