### 2024-02-11 c2b3734
Update web-app.md
typo
