### 2024-02-14 09611cb
V4 Update
Overall:
Things can now work in demo mode when the Dev folder is present
Added a way to test the TCP connection based on configured IP and ports
Enhancements to sidebar component
other minor enhancements

Automation:
Added the autoTipShape functionality and made it compatible with the existing autoInit page
