+++ scanbot/src/Scanbot.js
@@ -5,6 +5,7 @@ import "./styles/Scanbot.css";
 
 function Scanbot() {
 	const [configFound, setConfigFound] = useState(true)
+	const [isConnected, setIsConnected] = useState(true)
 
 	const CardRow = () => (
 		<div className="card-row">
@@ -18,8 +19,15 @@ function Scanbot() {
 		const hasConfig = await getResponse('/has_config')
 		setConfigFound(hasConfig)
 	}
+
+	const testConnection = async () => {
+		const hasConnection = await getResponse('/test_connection')
+		setIsConnected(hasConnection)
+	}
+
 	useEffect(() => {
 		checkConfig()
+		testConnection()
 	// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
 
@@ -29,6 +37,7 @@ function Scanbot() {
 				<header className="scanbot-header">
 					<h1>Scanbot</h1>
 				</header>
+				{isConnected ? <div/> : <p className='config-warning'><strong>Warning:</strong> Could not connect to Nanonis. Please check configured IP and ports and make sure Nanonis V5 is running.</p>}
 				{configFound ? <div/> : <p className='config-warning'><strong>Warning:</strong> A configuration file has not been found - using default parameters. Accept configuration to remove this warning.</p>}
 				<div className="scanbot-body">
 					<CardRow />
