+++ server/server.py
@@ -1,4 +1,5 @@
-from flask import Flask, request, send_from_directory, jsonify
+from flask import Flask, request, send_from_directory
+from flask_cors import CORS
 from scanbot_interface import scanbot_interface
 from scanbot_config import scanbot_config
 import numpy as np
@@ -7,23 +8,53 @@ import shutil
 from pathlib import Path
 from PIL import Image
 import base64
-import time
+import sys
 import pickle
-app = Flask(__name__)
+import global_
+import webbrowser
+from threading import Timer
+
+# pyinstaller --onefile --icon=..\scanbot\public\favicon.ico --add-data "..\scanbot\build;static" --name scanbot v4.x server.py
+app = Flask(__name__, static_url_path='')
+
+# Determine if we're running in a PyInstaller bundle and adjust paths
+if getattr(sys, 'frozen', False):
+    # If the app is run from a PyInstaller bundle
+    app.static_folder = os.path.join(sys._MEIPASS, "static")
+else:
+    # Otherwise, use the normal base directory
+    app.static_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'scanbot', 'build'))
+
+@app.route('/', defaults={'path': ''})
+@app.route('/<path:path>')
+def serve(path):
+    if path and os.path.exists(os.path.join(app.static_folder, path)):
+        return send_from_directory(app.static_folder, path)
+    else:
+        return send_from_directory(app.static_folder, 'index.html')
 
 @app.route('/has_config')
 def has_config():
     hasConfig = os.path.isfile("scanbot_config.ini")
     return {"status": hasConfig}, 200
 
+@app.route('/test_connection')
+def test_connection():
+    status = scanbot.testConnection()
+    return {"status": status}, 200
+
 @app.route('/reset_init')
 def reset_init():
     scanbot.scanbot.autoInitSet = False
     return {"status": "success"}, 200
 
+@app.route('/demo_init')
+def demo_init():
+    print("calling demo init",scanbot.autoInit(user_args=['-demo=1','-reactInit=1']))
+    return {"status": "success"}, 200
+
 @app.route('/is_auto_init')
 def is_auto_init():
-    print(scanbot.scanbot.autoInitSet)
     return {"status": scanbot.scanbot.autoInitSet}, 200
 
 @app.route('/check_hook', methods=['POST'])
@@ -87,7 +118,8 @@ def save_frame():
 
 @app.route ('/get_initialised_frame')
 def get_initialised_frame():
-    return send_from_directory('./autoinit', 'initialisation.png', as_attachment=True)
+    autoinit_dir = getDir('./autoinit')
+    return send_from_directory(autoinit_dir, 'initialisation.png', as_attachment=True)
 
 @app.route('/scanbot_config')
 def get_config():
@@ -135,19 +167,41 @@ def run_biasdep():
         pass
     return {"status": "success"}, 200
 
+@app.route('/run_autotipshape', methods=['POST'])
+def run_autotipshape():
+    userArgs = request.json['userArgs']
+    error = scanbot.autoTipShape(user_args=userArgs)
+    if(error):
+        return {"status": error}, 503
+    try:
+        shutil.rmtree('./temp')
+    except:
+        pass
+    return {"status": "success"}, 200
+
+@app.route('/get_imprint_score')
+def get_imprint_score():
+    files = sorted([file for file in Path('./temp').iterdir() if file.suffix == '.png'], key=os.path.getmtime)
+    latestFile = str(files[-1].name)
+    size = latestFile.split("size--")[1].split("_")[0]
+    symm = latestFile.split("symm--")[1].split(".png")[0]
+    running = global_.running.is_set()
+    return {"size": size,"sym": symm, "running": running}, 200
+    
+
 @app.route('/image_updates', methods=['POST'])
[…]
