+++ docs/web-app.md
@@ -11,7 +11,7 @@ This application is still being developed, with the following features already f
 
 ## Installation
 
-To use the web application, you can [download and run the .exe]([downloading and running the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2Fscanbot.zip?alt=media&token=ee1091ef-7b08-4ec4-903d-3892a0fbd7b0))
+To use the web application, you can [download and run the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2Fscanbot_v4.zip?alt=media&token=83f256ca-c22d-4818-b6e7-225870796380)
 , then navigate to [http://127.0.0.1:5000/](http://127.0.0.1:5000/).
 Alternatively, follow the normal installation and complete these additional steps:
 
