### 2024-02-19 f1bd065
Autotip shaping with full automation in web app
