+++ server/scanbot.py
@@ -66,9 +66,14 @@ class scanbot():
     xyMinF = 500                                                                # Frequency limits Dummy value - this gets overridden by config.
     
     autoInitSet  = False                                                        # Flag to indicate whether tip, sample, and clean metal locations have been initialised
-    
+    autoInitDemo = 0                                                            # 0 = live, 1 = demo
+
     surveyParams  = []                                                          # Last survey params
     survey2Params = []                                                          # Last survey2 params
+
+    currentAction = {"action":  "",                                             # Let the front-end know what Scanbot is currently doing
+                     "status":  "",
+                     "message": ""}
 ###############################################################################
 # Constructor
 ###############################################################################
@@ -142,7 +147,7 @@ class scanbot():
         iamauto : Flag when this function was called autonomously
 
         """
-        
+        self.currentAction["action"] = "survey"
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error):
             global_.running.clear()                                             # Free up the running flag
@@ -237,9 +242,11 @@ class scanbot():
                     
                 classificationHistory.append(classification)
                 
+                print("Image classification: ",classification)
+                
                 if(classification["tipShape"] == 1):
                     print("auto tip shaping initate!")
-                    global_.running.clear()
+                    # global_.running.clear()
                     callAutoTipShape = True
             
             _,scanData,_ = scan.FrameDataGrab(self.channel, 1)                  # Grab the data related to our focused channel
@@ -266,7 +273,8 @@ class scanbot():
                 self.interface.uploadToCloud(pklFile)                           # Send data to cloud database
             
             if(self.checkEventFlags()): break                                   # Check event flags
-        
+            if(callAutoTipShape): break
+
         if(stitch == 1 and not np.isnan(stitchedSurvey).all()):
             stitchFilepath = self.makePNG(stitchedSurvey,pngFilename = suffix + '_stitch.png',dpi=150*n, fit=False)
             self.interface.sendPNG(stitchFilepath,notify=False,message=message) # Send a png over zulip
@@ -282,6 +290,7 @@ class scanbot():
         if(iamauto): return callAutoTipShape
         
         if(callAutoTipShape):
+            global_.running.clear()                                             # Free up the running flag
             user_args = ['-run=survey', '-return=1', '-tipshape=1']
             self.interface.moveTipToClean(user_args=user_args)
         
@@ -292,6 +301,7 @@ class scanbot():
         each one.
         
         """
+        self.currentAction["action"] = "survey"
         
         self.survey2Params = [bias,n,startAt,suffix,xy,dx,px,sleepTime,stitch,survey_hk,classifier_hk,autotip, # Survey2 params
                               nx,ny,xStep,yStep,zStep,xyV,zV,xyF,zF]
@@ -387,6 +397,7 @@ class scanbot():
         Error message if something went wrong
 
         """
+        self.currentAction["action"] = "biasdep"
         NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
         if(connection_error):
             global_.running.clear()                                             # Free up the running flag
@@ -527,6 +538,7 @@ class scanbot():
                 
                    
         """
+        self.currentAction["action"] = "zdep"
[…]
