+++ docs/web-app.md
@@ -1,4 +1,4 @@
-## Scanbot Web App
+# Scanbot Web App
 
 Scanbot has been implemented as a web application using [React](https://react.dev/).
 This application is still being developed, with the following features already fully implemented:
@@ -9,9 +9,9 @@ This application is still being developed, with the following features already f
 
 ![WebApp](./appim/top-level.png)
 
-## Installation
+# Installation
 
-To use the web application, you can [download and run the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2Fscanbot_v4.zip?alt=media&token=83f256ca-c22d-4818-b6e7-225870796380)
+To use the web application, you can [download and run the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2FScanbot.zip?alt=media&token=25b7a4a5-b18a-49d5-b4bf-ae35973e7094)
 , then navigate to [http://127.0.0.1:5000/](http://127.0.0.1:5000/).
 Alternatively, follow the normal installation and complete these additional steps:
 
@@ -20,7 +20,7 @@ Alternatively, follow the normal installation and complete these additional step
 3. Start the server: navigate to ```~/scanbot/server/``` and run ```python server.py```
 4. Start the web app: navigate to ```~/scanbot/scanbot/``` and run ```npm start```
 
-## Configuration
+# Configuration
 From the landing page, you can access Scanbot's configuration.
 It is reccommended that you review the default settings before using any of the tools in Scanbot.
 Upon accepting the configuration, your scanbot_config.ini file will be saved automatically and your settings will be remembered.
@@ -28,7 +28,7 @@ These settings can be updated any time. <strong>For a detailed description about
 [configuration](../configuration).</strong>
 ![Config](./appim/configuration.png)
 
-## Data Acquisition
+# Data Acquisition
 From the landing page, you can access the data acquisition tools, including tools to help automate the STM.
 ![DataAcquisition](./appim/data-acquisition.png)
 
@@ -82,23 +82,33 @@ When the Bias Dependent function is selected, the following parameters can be co
 | Backwards time per line multiplier                   | Sets the backward speed of the tip during drift correction scans. (0.5 = half the speed as the forward speed)                                                               |
 | Suffix                                               | Text to append to the .sxm filenames for all images in this set                                                                                                             |
 
-## Automation
+# Automation
 
-Surveys can be fully automated by first initialising the tip, sample, and clean reference metal locations from the 'Automation' screen - simply allow the browser to use of the camera and follow the prompts.
-Once these locations are initialised, you will be able to answer 'Yes' to the ```Automated Tip Shaping?``` field.
-Under these cirumstances, Scanbot will autonomously shape the tip <strong>once it detects a five 'bad' images in a row</strong>.
-The process is described in more details [here](../automation/#intended-usage).
+From the Automation screen, you can initiate the following actions:
 
-Alternatively, you can trigger automated tip shaping manually. This action tells Scanbot to [automatically acquire a tip](../automation/#tip-preparation) suitable for imaging. The process is as follows:
+* Automated tip shaping
+* Tip, sample, and reference metal initialisation
+* Move tip to sample/metal
+
+## Automated Tip Shaping
+
+Scanbot can [automatically acquire a tip](../automation/#tip-preparation) suitable for imaging via its tip-shaping protocol. The process is as follows:
 
 1. An NxN grid is setup within the scannable region
 2. A small scan is acquired at the first location in the grid
 3. If the scanned region is flat/clean, it is deemed suitable for creating an imprint. If not, the scan frame is moved to the next location in the grid and the process repeats
[…]
