### 2024-02-21 #6028 @ziatdinovmax
@ceds92 I was expecting something more along the lines of [continuous integration workflows](https://docs.github.com/en/actions/automating-builds-and-tests/about-continuous-integration). Do I understand correctly that testing of the scanbot requires a connection with an actual nanonis controller?
