### 2024-02-21 #6028 @ceds92
@ziatdinovmax

That's correct, testing Scanbot requires a TCP connection with nanonis/nanonis simulator controller software.

Apologies, I'm not super familiar with CI and based on what I've just read, it will take me a while to fully implement it for Scanbot. I have however, finished the development of the web application (V4 branch) that is able to launch from a .exe which would simplify things a lot as no installation is required. If you would consider testing this using the nanonis simulator, I can upload a document detailing the specific tests related to claims made in the joss submission. I think it would take around 30-45 mins to run those specific tests manually.

If implementing CI is something you require, I'll attempt to do so but may not have time for a couple of weeks.

Thanks for your time.
