### 2024-02-22 #6028 @zhubonan
The manual steps for testing is OK, although automated tests are strongly included. @ziatdinovmax

However, JOSS does require that the manual test steps to be included in the documentation. @ceds92

One more thing, you don't have to have tests for the "full" functionality - people usually write unit tests to ensure the code correctness. These tests are usually for different functions which may not involved external software at all. For interaction with external software, one way to get around is to use mocks in the code, e.g. simulated responses from the external software.
