### 2024-02-28 8b5e6db
Deprecate browser mode and update testing steps with nanonis simulator
