+++ docs/commands.md
@@ -59,3 +59,83 @@ For a list of available commands on the fly, run ```help```, or for help with a 
 | auto_tip_shape                          | ```-n```: (int) Max number of tip shapes to perform. -1 means keep going until criterea met. Default: ```10```<br>```-wh```: (float) Size of the square scan frame when imaging the clean surface. Default: ```10e-9```<br>```-sym```: (float) Minimum circularity requirement. 0=don't care, 1=perfectly circle. Default: ```0.7```<br>```-size```: (float) Max size of the desired tip imprint in units of nm2. Default: ```1.2```<br>```-zQA```: (float) z-lift when performing a light tip shape to asses quality of tip (m). Default: ```-0.9e-9```<br>```-ztip```: (float) z-lift when performing a tip shape to alter the tip (m). Default: ```-1.5e-9```<br>```-st```: (int) Drift compensation time (s). Default: ```10```                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      | Routine that automatically prepares a nice tip. Assumes the substrate is a clean metal.                   | auto_tip_shape -sym=0.9 -size=2.1 |
 | auto_init                               | ```-light```: (int) Flag to turn the light on/off before/after initialisation. This uses hk_light.turn_on() and hk_light.turn_off() functions. 0=Don't, 1=Do. Default: ```0```<br>```-cameraPort```: (int) cv2 camera port - usually 0 for desktops or 1 for laptops with an inbuilt camera.. Default: ```0```<br>```-demo```: (int) Load in an mp4 recording of the tip moving instead of using live feed. Default: ```0```<br>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         | Initialise tip, sample, and clean metal positions using the camera feed                                   | auto_init -cameraPort=0           |
 | move_tip_to_sample<br>move_tip_to_clean | ```-light```: (int) Flag to turn the light on before moving and off after moving. This uses hk_light.turn_on() and hk_light.turn_off() functions. 0=Don't, 1=Do. Default: ```0```<br>```-cameraPort```: (int) cv2 camera port - usually 0 for desktops or 1 for laptops with an inbuilt camera. Default: ```0```<br>```-xStep```: (int) Number of motor steps in the X direction before updating tip position. More=Faster but might lose the tip. Default: ```100```<br>```-zStep```: (int) Number of motor steps to move in +Z (upwards) before updating tip position. More=Faster but might lose the tip. Default: ```250```<br>```-xV```: (float) Piezo voltage when moving motor steps in x direction. Default: ```130```<br>```-zV```: (float) Piezo voltage when moving motor steps in z direction. Default: ```180```<br>```-xF```: (float) Piezo frequency when moving motor steps in x direction. Default: ```1100```<br>```-zF```: (float) Piezo frequency when moving motor steps in z direction. Default: ```1100```<br>```-approach```: (int) Approach when tip reaches target. 0=No,1=Yes. Default: ```1```<br>```-demo```: (int) Load in an mp4 recording of the tip moving instead of using live feed. Default: ```0``` | After auto_init has been run, automatically move the tip to sample/clean metal locations using the camera | move_tip_to_sample -light=1       |
+
+
+# Intended Usage
[…]
