+++ docs/automation.md
@@ -51,84 +51,4 @@ the spectra obtained using the tip on a clean metal surface closely corresponds 
 Thanks to [Wang, et al.](https://pubs.acs.org/doi/pdf/10.1021/acs.jpca.0c10731), the groundwork for this has already been done so integrating this feature into Scanbot should be trivial!
 In addition to this, automated STS commands are also in development.
 
-<br>
-
-## Intended Usage
-This section describes how the automation aspect of Scanbot can be used for supervised or unsupervised operation. Note that you can still take advantage of Scanbot's
-[data acquisition commands](../commands/#data-acquisition) without automating the entire process.
-
-### 1. Setup
-These steps must be carried out before Scanbot is able to automatically move the tip between the sample and the clean metal.
-
-1. Making sure the STM tip, sample of interest, and clean metal surface are in clear view of the camera.
-2. Run the ```auto_init``` command. This will initialise the locations of the STM tip, sample, and clean metal.
-    - Drive the tip completely outside the view of the camera, then click within the picture to set the background frame.
-    - Drive the tip back into view of the camera, in close proximity to the sample.
-    - Place a marker near the apex of the tip.
-    - Place a marker at a safe distance above the clean metal.
-    - Place a marker at a safe distance above the sample.
-    - **Note: Close any other applications that use the camera as they might prevent Scanbot from accessing it.**
-    - **Note2: If the course piezos are moved manually at any point, Scanbot will have incorrect information about the location of the tip and this process must be repeated. Same goes for moving the camera.**
-
-### 2. End-to-End Supervised Test
-**It is advised to go through these steps, especially steps 1 and 3, each time ```auto_init``` is run.**
-
-1. Run the ```move_tip_to_clean``` command. This will move the tip to the location you set above the clean metal and begin an auto approach.
-    - **Configure the correct piezo voltages, frequency, and number of steps for your machine** (see [commands](commands.md) or run ```help move_tip_to_clean```).
-    - Observe the camera feed and ensure the tip is tracked correctly. If not, run ```stop```, then rerun ```auto_init``` after refocusing the camera.
-    - Note: run the ```stop``` command at any time to abort.
-2. Once the tip has approached, run ```auto_tip_shape``` with the appropriate inputs for your system (e.g. metal used, etc.).
-    - This function may take some time, for testing purposes run the ```stop``` command once you've [verified it's working correctly](#auto-tip-shaping).
-3. Next run the ```move_tip_to_sample``` command.
-    - Observe the camera feed and ensure the tip is tracked correctly.
-4. Once the tip has approached, start a survey using the ```survey``` or ```survey2``` command.
-
-### 3. Supervised Operation Example
-1. After setup, run the ```move_tip_to_sample``` command.
-2. Upon approach, start a ```survey``` leaving ```-autotip=0```.
-3. Periodically monitor the scans stream (for Zulip users, each scan will be sent through the steam the survey was started from).
-4. If the tip requires refinement, run ```stop``` followed by ```move_tip_to_clean``` with the options ```-tipshape=1```, ```-return=1```, and ```-run=survey```
-    - Here, Scanbot will automatically move the tip to the clean metal location set in ```auto_init```
-    - With ```-tipshape=1```, the auto tip shaping process will begin after approaching
-    - With ```-return=1```, Scanbot will move the tip back to the sample location set in ```auto_init```
-    - With ```-run=survey```, Scanbot will begin a survey with the same options as the previous survey
-    - **Note: If the function passed to the -run option has not been run previously, the default parameters will be used - check them to ensure they are compatible with your system in this case.**
-
-### 4. Fully Automated Operation
-1. After setup, run the ```survey``` or ```survey2``` command with the ```-autotip=1``` option.
[…]
