### 2024-02-28 #6028 @ceds92
Ok, thanks for the clarification @zhubonan. I will update the documentation to reflect the manual testing steps with the simulator. I will release it by the end of the week and update @ziatdinovmax when it's ready.

Thanks for your patience.
