+++ pyproject.toml
@@ -15,7 +15,7 @@ dependencies = [
     'numpy',
     'matplotlib',
     'scipy',
-    'opencv-python < 4.8.0',
+    'opencv-python < 4.9.0.81',
     'zulip',
     'panel',
     'flask',
