+++ docs/web-app-test.md
@@ -14,7 +14,7 @@ Completing all of these tests will take approximately one hour.
 
 ### Nanonis V5 Simulator
 1. Open the Nanonis V5 Simulator
-2. In the 'Current' module, change the Gain to LN 10^9
+2. In the 'Current' module, change the Gain to LN 10^9. <strong>If your version of the simulator does not have this setting you can ignore this step.</strong>
 <br>![CurrenModule](./testing/gain109.png)<br>
 3. In the Z-Controller module, turn on the Controller. Ensure the tip is 'tunnelling' at the correct setpoint.
 <br>
@@ -30,7 +30,9 @@ Completing all of these tests will take approximately one hour.
 <br>
 ![Scanbot-terminal](./testing/scanbot-terminal.png)
 <br>
-2. Scanbot should then open in a new browser tab. If it does not, then navigate to [http://127.0.0.1:5000](http://127.0.0.1:5000)
+2. Scanbot should then open in a new browser tab. If it does not, then navigate to [http://127.0.0.1:5000](http://127.0.0.1:5000).
+When running Scanbot for the first time, you will see a warning letting you know that you have not yet reviewed the configuration.
+You may receive an additional warning message if Scanbot is unable to communicate to Nanonis. This will be solved in the next step.
 <br>
 ![landing-page](./testing/top-level.png)
 <br>
@@ -46,6 +48,7 @@ Completing all of these tests will take approximately one hour.
 ![configuration-page](./testing/configuration-page.png)
 <br>
 4. When running Scanbot using the simulator, the default configuration is ok. Accept the configuration. This will save the scanbot_config.ini file and your settings will be remembered.
+Accepting the configuration will also stop the warning message from being displayed.
 
 ## Sample Surveying (10 min)
 1. From the Scanbot landing page, navigate to 'Data Acquisition' => 'Survey'.
@@ -54,6 +57,7 @@ Alternatively, you can hover over any field to obtain its description. In summar
 <br>
     * The Fine Grid Parameters control the size and spacing of each survey grid.
     * The Coarse Grid Parameters control the number of surveys to acquire (at different macroscopic locations) and the spacing between them.
+      <strong>Note: The motor controller in the Nanonis simulator is not sufficient for testing the coarse movement of the STM tip, thus please leave the coarse grid size to 1x1.</strong>
     * The Scan Parameters configure the settings for individual scans.
 <br>
 <strong>Note: At this stage, you will not be able to answer "Yes" to "Automated Tip Shaping?". We'll go through that later.</strong>
@@ -83,6 +87,8 @@ In Summary:
 ![tip-shaper-module](./testing/tip-shaper-module.png)
 <br>
 3. In Nanonis, from 'Graphs', open the 'Signal Chart'. This can be used to monitor Scanbot's tip shaping actions (as well as looking at the parameters in the Tip Shaper module).
+Select the 'Current (A)' and 'Z (m)' channels. You can adjust the 'Averaging' slider to slow down the rate at which the curves move through the window.
+In the figure below, the tip was at an intial height of -2 nm, then moved 3 nm into the surface and retracted 8 nm before turning the feedback controller on. During this time, the current spiked to its maximum value of 10 nA.
 <br>
 ![tip-shaping](./testing/tip-shaping.png)
 <br>
@@ -95,7 +101,7 @@ In Summary:
     * <strong>The last imprint in the demo data has an imprint size of 1.55 nm<sup>2</sup> and a circularity of 0.83.</strong>
         Set the parameters for desired size and circularity appropriate to these numbers.
 5. Click 'Start'
-6. You may find that the scan frame repeatedly moves before a scan is able to complete.
[…]
