### 2024-02-29 7c783b2
Minor changes
Fix so motor doesn't cause crash in demo mode.
Make the web app back button white
