+++ docs/index.md
@@ -25,7 +25,7 @@ Scanbot is a collection of several automated STM and nc-AFM data acquisition com
     <summary>Web App</summary>
     Scanbot has been implemented as a web application using [React](https://react.dev/).
     
-    On Windows, the easiest way to use it is by [downloading and running the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2FScanbot_V4.zip?alt=media&token=0dfed896-3df8-4a44-81e5-bf0b126181d9).
+    On Windows, the easiest way to use it is by [downloading and running the .exe](https://firebasestorage.googleapis.com/v0/b/scanbot-46390.appspot.com/o/scanbot-react%2FScanbot_V4.zip?alt=media&token=1c627e79-422e-41e9-9d85-dd12835c3386).
     
     Alternatively, you can install Scanbot from its source:
 
