### 2024-03-01 36da852
Fix crash caused by motor module when using the simulator
