### 2024-03-01 #6028 @ceds92
I've updated the documentation with the manual test steps @ziatdinovmax. Everything needed for testing can be found here:
https://new-horizons-spm.github.io/scanbot/web-app-test/

Let me know if there's anything else 😃
