### 2024-03-12 #5432 @ianhi
Am I correct in interpreting the below as saying that the free mimea substitutes for the nanonis V5?

> Testing this software requires Spec's Nanonis V5 which is a licensed product. Further, it may be challenging for reviewers to find time in their schedules to pause their research and test this software on a live system (STM). However, this Scanbot can be tested on the STM simulator V5 software provided by Specs [here](https://www.specs-group.com/nanonis/products/mimea/). All functionality except for Scanbot's control over the STM tip via the motors + camera feed can be tested using this simulator.

I unfortunately have no experience with STM. That said if no one with expertise is available and it is possible to test without access to the licensed software/STM then I would be wiling to review. However, I would not be able to start until next week.
