### 2024-03-13 #5432 @ceds92
Hi @ianhi, thanks so much for agreeing! Yes, the software can be tested with the freely available STM simulator. I hope I have included enough detail in the documentation such that a non-STM'r can still download and test the software. Otherwise please ask if you have any questions.

https://new-horizons-spm.github.io/scanbot/web-app-test/

Your time is very much appreciated!
