### 2024-03-14 4759735
Fix packaging
Scanbot can now be installed with 'pip install .', then run with 'scanbot' command
