+++ pyproject.toml
@@ -1,26 +1,38 @@
-[build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
+[tool.poetry]
 name = "scanbot"
 version = "1.0.0"
 description = "Collection of automated STM and nc-AFM commands compatible with Nanonis V5 SPM Controller"
-authors = [{name="Julian Ceddia", email="jdceddia@gmail.com"}]
+authors = ["Julian Ceddia <jdceddia@gmail.com>"]
 readme = "README.md"
-requires-python = ">=3.7"
-license = {text = "MIT"}
-keywords = ["stm","automation","python","scanning tunnelling microscopy","sts","scanbot"]
-dependencies = [
-    'numpy',
-    'matplotlib',
-    'scipy',
-    'opencv-python < 4.9.0.81',
-    'zulip',
-    'panel',
-    'flask',
-    'flask-cors',
-]
+license = "MIT"
+keywords = ["stm", "automation", "python", "scanning tunnelling microscopy", "sts", "scanbot"]
+packages = [
+            {include = "scanbot/server"},
+            ]
+include = [
+            "scanbot/App/build/**/*",
+            "scanbot/App/public/**/*",
+            "scanbot/App/src/**/*",
+            "scanbot/Dev/**/*",
+            ]
 
-[tool.setuptools]
-packages = ["scanbot"]
+[tool.poetry.dependencies]
+python = "^3.9"
+numpy = "*"
+matplotlib = "*"
+scipy = "*"
+"opencv-python" = "<4.9.0.81"
+zulip = "*"
+panel = "*"
+flask = "*"
+flask-cors = "*"
+
+[tool.poetry.dev-dependencies]
+pytest = "*"  # Add other development dependencies as needed
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+scanbot = "scanbot.server:app_"

[…]
