### 2024-03-14 72bd74b
Remove version arg
Moved version to server.py
