+++ scanbot/server/scanbot_interface.py
@@ -959,9 +959,6 @@ class scanbot_interface(object):
 handler_class = scanbot_interface                                                   # Used by zulip-run-bot
 
 finish = False
-if('--version' in sys.argv):
-    print("Scanbot v4")
-    finish = True
     
 if('-z' in sys.argv and not finish):
     rcfile = ''
