+++ .github/workflows/windows-install.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Install node_modules and build the app
       run: |
         cd scanbot/App
-        npm install
+        npm install --loglevel=error  # Only log errors, ignore warnings
         npm run build
 
     - name: Install the main application
