### 2024-03-15 e0cf2ff
Initial pytest for server (WIP)
