+++ scanbot/server/server.py
@@ -18,9 +18,6 @@ import argparse
 parser = argparse.ArgumentParser()
 parser.add_argument('--version', action='version', version='scanbot 4.1.0', help='show the version number and exit')
 args = parser.parse_args()
-if args.version:
-    print(f"scanbot {args.version}")
-    sys.exit()
 
 # pyinstaller --onefile --icon=..\App\public\favicon.ico --add-data "..\App\build;static" --name scanbot_v4.1 server.py
 app = Flask(__name__, static_url_path='')
@@ -29,14 +26,14 @@ app = Flask(__name__, static_url_path='')
 if getattr(sys, 'frozen', False):
     # If the app is run from a PyInstaller bundle
     app.static_folder = os.path.join(sys._MEIPASS, "static")
-    module_dir = "./"
+    app.module_dir = "./"
 else:
     # Otherwise, use the normal base directory
     app.static_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'App', 'build'))
-    module_dir = str(os.path.dirname(os.path.abspath(__file__))).replace('\\','/')
-    if(not module_dir.endswith('/')): module_dir += '/'
+    app.module_dir = str(os.path.dirname(os.path.abspath(__file__))).replace('\\','/')
+    if(not app.module_dir.endswith('/')): app.module_dir += '/'
 
-scanbot = scanbot_interface(run_mode='react',module_dir=module_dir)
+scanbot = scanbot_interface(run_mode='react',module_dir=app.module_dir)
 
 @app.route('/', defaults={'path': ''})
 @app.route('/<path:path>')
@@ -48,7 +45,7 @@ def serve(path):
 
 @app.route('/has_config')
 def has_config():
-    hasConfig = os.path.isfile(module_dir + "scanbot_config.ini")
+    hasConfig = os.path.isfile(app.module_dir + "scanbot_config.ini")
     return {"status": hasConfig}, 200
 
 @app.route('/test_connection')
@@ -90,11 +87,11 @@ def save_frame():
     initialFrame = data['initialFrame']
     tipInFrame   = data['tipInFrame']
 
-    Path(module_dir + 'autoinit').mkdir(parents=True, exist_ok=True)
+    Path(app.module_dir + 'autoinit').mkdir(parents=True, exist_ok=True)
 
     header, encoded = initialFrame.split(",", 1)
     binary_data = base64.b64decode(encoded)
-    file_path = module_dir + 'autoinit/initialFrame.png'
+    file_path = app.module_dir + 'autoinit/initialFrame.png'
     with open(file_path, 'wb') as file:
         file.write(binary_data)
     
@@ -104,7 +101,7 @@ def save_frame():
         
     header, encoded = tipInFrame.split(",", 1)
     binary_data = base64.b64decode(encoded)
-    file_path = module_dir + 'autoinit/tipInFrame.png'
+    file_path = app.module_dir + 'autoinit/tipInFrame.png'
     with open(file_path, 'wb') as file:
         file.write(binary_data)
     
@@ -120,7 +117,7 @@ def save_frame():
         'tipInFrame'      : np.array(tipInFrame)
     }
 
-    pickle.dump(pk,open(module_dir + 'autoinit/autoinit.pk','wb'))
+    pickle.dump(pk,open(app.module_dir + 'autoinit/autoinit.pk','wb'))
 
     userArgs = ['-reactInit=1']
     error = scanbot.autoInit(user_args=userArgs)
@@ -131,7 +128,7 @@ def save_frame():
 
 @app.route ('/get_initialised_frame')
 def get_initialised_frame():
-    autoinit_dir = getDir(module_dir + 'autoinit')
[…]
