+++ pyproject.toml
@@ -26,6 +26,7 @@ zulip = "*"
 panel = "*"
 flask = "*"
 flask-cors = "*"
+pytest = "*"
 
 [tool.poetry.dev-dependencies]
 pytest = "*"  # Add other development dependencies as needed
