### 2024-03-17 d254a20
Add test for demo_init
Test the demo tip initialisation procedure
