+++ docs/commands.md
@@ -1,3 +1,12 @@
+<!-- Google tag (gtag.js) -->
+<script async src="https://www.googletagmanager.com/gtag/js?id=G-6MK4DRHXWM"></script>
+<script>
+  window.dataLayer = window.dataLayer || [];
+  function gtag(){dataLayer.push(arguments);}
+  gtag('js', new Date());
+
+  gtag('config', 'G-6MK4DRHXWM');
+</script>
 # Commands
 This page documents the available commands and how to use them when running Scanbot via Zulip or in a terminal.
 For a list of available commands on the fly, run ```help```, or for help with a specific command, run ```help <command_name>```
