### 2024-03-17 #6028 @ceds92
Hi @zhubonan and @ziatdinovmax,

This week I started implementing the [automated tests](https://github.com/New-Horizons-SPM/scanbot/actions/workflows/windows-install.yml). So far this includes installation and about 50% of the web app server endpoints. I didn't get to spend as much time on it as I'd hoped for this week but plan to implement the remaining unit tests next week.

Anyway, in doing this I had to refine the installation procedure so just giving you a heads up about that. [The documentation has the most recent installation steps](https://new-horizons-spm.github.io/scanbot/web-app/#installation)
