### 2024-03-20 f9be98d
nanonisTCP is pip installable
no need for manual installation of dependent package nanonisTCP since it became installable with pip.
added nanonisTCP to pyproject.toml
