### 2024-03-22 984acf1
Make terminal and zulip modes compatible with the new installation method
Run Scanbot in a terminal with scanbot -c
Run Scanbot using zulip with scanbot -z

Also bug fix biasdep front end - initial bias and final bias fields
