### 2024-03-22 1fa53c8
Update to reflect v4.2.0
Terminal is now run by scanbot -c
Zulip is now run by scanbot -z

Update new download link
