+++ scanbot/server/test_server.py
@@ -6,6 +6,9 @@ import threading
 import time
 import cv2
 import os
+from PIL import Image
+import base64
+import io
 
 @pytest.fixture
 def client(tmp_path):
@@ -14,11 +17,17 @@ def client(tmp_path):
         yield client
 
 def test_homepage(client):
+    """
+    Test that the server is able to provide the homepage
+    """
     response = client.get('/')
     homepage_found = response.data.decode().startswith('<!doctype html>')
     assert homepage_found == True
 
 def test_has_config(client, tmp_path):
+    """
+    Test that the config file can be found
+    """
     fake_file = tmp_path / "scanbot_config.ini"
     fake_file.write_text("test")
     
@@ -29,6 +38,9 @@ def test_has_config(client, tmp_path):
     assert file_found == True
 
 def test_has_no_config(client):
+    """
+    Test that the config file can't be found
+    """
     response = client.get('/has_config')
     json_data = response.json
     file_found = json_data['status']
@@ -36,10 +48,27 @@ def test_has_no_config(client):
     assert file_found == False
 
 def test_test_connection(client):
-    response = client.get('/test_connection')
-    assert response.status_code == 200          # Hard to test anything else here since nanonis needs to be open
+    """
+    Test that the connection to nanonis can be verified
+    """
+    with patch('scanbot.server.scanbot.nanonisTCP') as mock_nanonisTCP, \
+         patch('scanbot.server.scanbot.Bias') as mock_Bias:
+         
+        mock_nanonisTCP.return_value.close_connection.return_value = None
+
+        mock_Bias.return_value.Get.return_value = 0 # Scanbot tests the connection by attempting to read the bias value in nanonis. Give it something
+
+        response = client.get('/test_connection')
+        assert response.status_code == 200
+        
+        json_data = response.json
+        connected = json_data['status']
+        assert connected
 
 def test_reset_init(client):
+    """
+    Test that the tip initiliasation can be reset
+    """
     response = client.get('/reset_init')
     json_data = response.json
     success = json_data['status']
@@ -47,11 +76,12 @@ def test_reset_init(client):
     assert response.status_code == 200
     assert success == "success"
 
-# %%
-# Test demo initialisation
-# 
-# First test that /is_auto_init returns False (since initialisation has not been run yet)
+# Test initialisation
+
 def test_is_not_auto_init(client):
+    """
+    Test to see if initialisation is not complete
+    """
     response     = client.get('/is_auto_init')
     json_data    = response.json
     is_auto_init = json_data['status']
@@ -61,6 +91,10 @@ def test_is_not_auto_init(client):
 
 # Next, go through the demo initialisation then retest to see of is_auto_init == True
 def wait_for_mouse_callback(mock_setMouseCallback):
+    """
+    This function grabs the mouse_callback function once it has been set by scanbot
+    It waits a maximum of 5s (should be very quick) before timing out.
+    """
     i = 0
     timeout = 5
     while mock_setMouseCallback.call_args is None:
@@ -71,6 +105,10 @@ def wait_for_mouse_callback(mock_setMouseCallback):
     return mock_setMouseCallback.call_args[0][1]
 
 def simulate_autoInit(mock_setMouseCallback, mock_waitKey):
+    """
+    This function simulates the initialisation for automatic tip shaping.
[…]
