### 2024-03-23 #6028 @ceds92
Hi @zhubonan and @ziatdinovmax,

I have completed the [automated tests](https://github.com/New-Horizons-SPM/scanbot/actions/workflows/windows-install.yml). Please let me know if there is anything else.
