### 2024-03-25 c4a566a
External download link
Redirect to latest download link - saves having to update links here each time a new .exe is released
