### 2024-04-02 #6028 @ziatdinovmax
> I have completed the [automated tests](https://github.com/New-Horizons-SPM/scanbot/actions/workflows/windows-install.yml). Please let me know if there is anything else.

Thanks! It is generally advised to have tests in a dedicated directory that is separate from the main package code, e.g.

```
mypackage/
    __init__.py
    module1.py
    module2.py
tests/
    test_module1.py
    test_module2.py
```

Other than that all looks good.
