### 2024-04-09 #6028 @ziatdinovmax
@ceds92 - Thanks!
One minor remaining boilerplate task from the JOSS checklist is "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".

Could you please add these to ReadMe and/or Documentation?
