### 2024-04-15 #6028 @ceds92
> @ceds92 - Thanks! One minor remaining boilerplate task from the JOSS checklist is "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".
>
> Could you please add these to ReadMe and/or Documentation?

Hey @ziatdinovmax, I have updated the [documentation](https://new-horizons-spm.github.io/scanbot/#contributing) to address this. If there's anything else, let me know. (apologies for the delay, I have been very busy in the lab)
