### 2024-04-18 a02fbc1
Implemented working drift-corrected STS Grids in data acquisition
To Do:

1) Make nice image for the STS-Grid card
2) Update documentation
3) Make the 'path' user-input work (save gridData in specified path instead of the working directory)
4) Update pytest to automate testing of sts-grid
5_ Comment code
