+++ scanbot/server/scanbot.py
@@ -17,6 +17,8 @@ from nanonisTCP.Current import Current
 from nanonisTCP.FolMe import FolMe
 from nanonisTCP.Marks import Marks
 from nanonisTCP.TipShaper import TipShaper
+from nanonisTCP.Pattern import Pattern
+from nanonisTCP.BiasSpectr import BiasSpectr
 
 from scanbot.server import global_
 from scanbot.server import utilities
@@ -511,7 +513,218 @@ class scanbot():
         self.disconnect(NTCP)                                                   # Close the TCP connection
         global_.running.clear()                                                 # Free up the running flag
             
-            
+
+    def getGrid(self,gridFrame):
+        nx,ny,ox,oy,w,h,gridAngle = gridFrame
+
+        dx  = w/nx
+        xx  = np.arange(nx) - nx/2 + 0.5*((nx+1)%2)
+        xx *= dx
+        xx += ox + dx/2
+        
+        dy  = h/ny
+        yy  = np.arange(ny) - ny/2 + 0.5*((ny+1)%2)
+        yy *= dy
+        yy += oy + dy/2
+
+        return xx,yy
+    
+    def stsGrid(self, params, message=""):
+        """
+        This function obtains an STS grid with drift correction every n points.
+        The tip must be inside the scan frame to start
+        The grid centre must also be inside the scan frame to start
+
+        Process:
+            1. The existing grid in Nanonis is loaded
+            2. An image is taken in the upward scan direction as a reference to measure drift.
+            3. The Tip moves to the first location in the grid and STS setpoint is applied
+            4. STS is acquired
+            5. The tip moves to the next point in the grid
+            6. Repeat steps 4 and 5 until n spectra have been acquired
+            7. Take another image in the upward direction and compare with the first to measure the amount of drift
+            8. Move the scan frame and the grid centre according to detected drift
+            9. Repeat steps 5 to 9 until the grid completes
+        """
+        Vset = params['-Vset']                                                  # Bias at which the setpoint for STS is taken
+        Iset = params['-Iset']                                                  # Setpoint current for STS
+        Vmov = params['-Vmov']                                                  # Bias while the tip is moving between grid points
+        Imov = params['-Imov']                                                  # Setpoint current while the tip is moving between grid points
+        VDC  = params['-VDC']                                                   # Bias for drift correction images
+        IDC  = params['-IDC']                                                   # Bias for drift correction images
+        NDC  = params['-NDC']                                                   # Take a drift correction frame after every NDC STS acquisitions
+        pxdc = params['-pxdc']
+        tdc  = params['-tdc']
+        tbdc = params['-tbdc']
+        path = params['-path']                                                  # Save the pickled grid array at this path
+        sufx = params['-s']
+
+        gridType = "grid"                                                       # Only grid is support at the moment. Support for cloud and line will be added later
+
+        self.currentAction["action"] = "stsgrid"
+        NTCP,connection_error = self.connect()                                  # Connect to nanonis via TCP
+        if(connection_error):
+            global_.running.clear()                                             # Free up the running flag
+            return connection_error                                             # Return error message if there was a problem connecting        
+        
+        scanModule = Scan(NTCP)
+        biasModule = Bias(NTCP)
+        zController = ZController(NTCP)
+        folme = FolMe(NTCP)
+        pattern = Pattern(NTCP)
+        bspec   = BiasSpectr(NTCP)
+        
+        tipPos    = folme.XYPosGet(Wait_for_newest_data=1)
[…]
