### 2024-04-18 #5432 @zhubonan
Hi @KoenImdea would you be able to help review this JOSS submission? It is about automating scanning tunnelling microscope and help for things like tip preparation.

We are currently short for reviewers. There is already one reviewer who have completed the review and recommanded for publication. However, we do need to have two reviews in order to accept the work. Your contribution would be really appreciated.

The manuscript can be viewed here 👉 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.06028/joss.06028/10.21105.joss.06028.pdf)
Software repository 👉 https://github.com/New-Horizons-SPM/scanbot
The JOSS reviewer policy and reviewer guideline for your reference 👉 [Reviewer Guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance.
