### 2024-04-19 85f82a1
Update documentation for STS Grid
