### 2024-04-21 e5933ae
STS grids info in index.md
