### 2024-04-22 #5432 @zhubonan
Thanks, @KoenImdea, glad to know you will be able to review the submission!

We have documentation about how to carry out the review. It basically involves: 1. generating a checklist. 2. going through the items in the checklist. You will need to download, install and test the code. If you have any comment or problem, you can directly raise an GitHub issue in the code's repository and mention the JOSS review issue's link. You can also leave comments and ask questions in the review issue directly.

Once you are satisfied with the submission  and all the boxes are ticked, just leave a comment for recommending acceptance in the review issue and ping me.
