### 2024-04-22 #6028 @zhubonan
Hi @ianhi to generate the checklist, please leave a comment containing

only. I don't think the editorial bot will be able to pick it up if the command is embedded inside a comment.
