### 2024-04-22 #6028 @ceds92
@KoenImdea, in case it's not obvious in the history of this thread, I should also mention that to test the claims in the paper you can follow the steps in this section of the [documentation](https://new-horizons-spm.github.io/scanbot/web-app/) using the nanonis simulator.
