### 2024-04-26 b715663
Bug fix when autosave is off and Z-Channel not selected
Autosave is now enabled automatically. The Z-Channel is also enabled automatically
