+++ scanbot/server/scanbot.py
@@ -201,10 +201,14 @@ class scanbot():
         if(stitch == 1):
             _,_,px,lines = scan.BufferGet()
             stitchedSurvey = np.zeros((lines*n,px*n))*np.nan
-        
+
         callAutoTipShape = False
         classificationHistory = []
         for idx,frame in enumerate(frames):
+            seriesName = scan.PropsGet()[3]
+            scan.PropsSet(continuous_scan=2,bouncy_scan=2,autosave=1,series_name=seriesName)
+            self.plotChannel(c=-1,a=14)
+
             if(idx < startAt-1): continue
             
             self.interface.sendReply('Running scan ' + str(idx + 1) + '/' + str(n**2),message=message) # Send a message that the next scan is starting
@@ -449,11 +453,15 @@ class scanbot():
         dy    = scanFrame[3]/lxdc
         dxy   = np.array([dx,dy])
         ox,oy = np.array([0,0])
-        
+
         GIF       = []
         biasList  = np.linspace(bi,bf,nb)
         initialDC = np.zeros((pxdc,pxdc))
         for idx,bias in enumerate(biasList):
+            seriesName = scanModule.PropsGet()[3]
+            scanModule.PropsSet(continuous_scan=2,bouncy_scan=2,autosave=1,series_name=seriesName)
+            self.plotChannel(c=-1,a=14)
+
             self.interface.sendReply("Scan " + str(idx+1) + "/" + str(nb))
             if(abs(dcbias) > 0):                                                # If drift correction is turned on, take a drift correction image
                 time.sleep(0.25)
@@ -613,9 +621,10 @@ class scanbot():
         if(IDC == 0):  IDC = Iset
         if(Vmov == 0): Vmov = Vset
         if(Imov == 0): Imov = Iset
-        
+
         seriesName = scanModule.PropsGet()[3]
         scanModule.PropsSet(continuous_scan=2,bouncy_scan=2,autosave=1,series_name=seriesName)
+        self.plotChannel(c=-1,a=14)
 
         pxdc  = scanModule.BufferGet()[2]
         dx    = scanFrame[2]/pxdc
@@ -657,7 +666,9 @@ class scanbot():
                     zController.SetpntSet(IDC)
                     time.sleep(0.1)
                     self.rampBias(NTCP, VDC)
-                    if(self.checkEventFlags()): break                               # Check event flags
+                    if(self.checkEventFlags()):
+                        stop = True
+                        break                                                       # Check event flags
                     time.sleep(0.25)
                     
                     basename_dc = "scanbot-DC-stsgrid"
@@ -688,7 +699,7 @@ class scanbot():
                     gridFrame[3] -= oy
                     pattern.GridSet(*gridFrame)
                     totalDrift = totalDrift - np.array([ox,oy])
-
+                
                 zController.SetpntSet(Imov)
                 self.rampBias(NTCP,Vmov)
                 time.sleep(0.1)
@@ -714,7 +725,9 @@ class scanbot():
                 saveDict = {"data" : gridData}
                 pickle.dump(saveDict,open(saveFilename,'wb'))
                 
-                if(self.checkEventFlags()): break                               # Check event flags
+                if(self.checkEventFlags()):
+                    stop = True
+                    break                                                       # Check event flags
 
                 count += 1
 
@@ -867,6 +880,10 @@ class scanbot():
         completionTime = dt.now() + timedelta(seconds=eta)
         self.interface.sendReply("Starting zdep.. ETA: " + str(completionTime))
         for idx,dz in enumerate(dzList):
+            seriesName = scanModule.PropsGet()[3]
[…]
