### 2024-04-26 #6028 @ceds92
Hi @KoenImdea,

Thanks so much for your detailed feedback! I am sorry and I hope that this bug has not wasted too much of your time.

Please see attached for my responses to your comments. I guess the main points are:

1) I was able to reproduce the errors you're encountering and they all stem from a case where autosave is not turned on and the Z channel is not selected. I have just pushed through a fix for this. Now, they are both automatically enabled.

2) I will update all of the comments relating to documentation at some point next week

Thanks again for looking into this!

[reply to Koen.docx](https://github.com/openjournals/joss-reviews/files/15127700/reply.to.Koen.docx)
