### 2024-04-26 #6028 @KoenImdea
Hi @ceds92,

Oh yes. As my last project with the Demo was without feedback I did not have the Z channel selected indeed. I did make sure that auto-save was enabled. I still believe that forcing it is better, and thank you for your change.

I still get the same error with the Z channel enabled however. I think this is due to how you grab the Z channel, in combination with some undocumented features of Nanonis.
From my console, output of your program:
>  File "D:\Koen\Python36\Lib\site-packages\scanbot\server\scanbot.py", line 230, in survey
    _,scanData,_ = scan.FrameDataGrab(14, 1)                            # Grab the data within the scan frame. Channel 14 is . 1 is forward data direction

Now from the sxm result-file:
> :DATA_INFO:
	Channel	Name	Unit	Direction	Calibration	Offset
	30	Z	m	both	7.500E-8	0.000E+0
	0	Current	A	both	1.000E-9	0.000E+0

It seems that Nanonis is not consistent with its channel numbers for even the basic topography. You expect topography Z to have number 14, in my case it is 30. Which I also never would have guessed if I had not talked with another lab from the other size of the world. In fact, it seems that I have no channel numbers in the range [8 - 23].
![V5e signal manager](https://github.com/openjournals/joss-reviews/assets/51089425/03b9f925-96b7-46f7-84b0-3cf06ca40356)

My guess is that they will defend it as a feature highlighting the flexibility of their equipment. I am unsure if this change in numbering has to do with us having the V5e version instead of the V5.

As I do not know your code, I do not know if it is feasible to look for the Name of the Z channel instead of the number? If not, you can think about adding this as a parameter in the configuration?
