### 2024-04-26 #6028 @ceds92
Hi @KoenImdea,

Ah damn ok that's annoying, I guess it's good that you're testing with a different version of Nanonis because my aim for this project was for other groups to be able to use it if they like it.

Thanks for your suggestion, I will see if I can download that version of Nanonis and have a look into this over the next couple of days.

I will keep you posted and let you know when I figure it out.

Thanks!
