### 2024-04-26 #6028 @ceds92
Hi @KoenImdea,

I contacted specs but unfortunately I am not able to install 5e with my licence. However, I checked one of our sxm files and in data_info, I see that Z is on channel 14, not 30 like yours. So it seems like the tcp command to get frame data differs between the two versions.

In version 5, channel refers to the 'slot' number.

In version 5e, channel refers to the 'RT Index'.

This is how it seems (I've sent an email to specs for them to confirm). I will take your suggestion and try to find the channel number based on the channel name.
