### 2024-04-26 #6028 @KoenImdea
Hi @ceds92 ,
Thanks for looking into this. In my experience the manufacturers of SPM electronics for some reason like to have a few of these undocumented surprises.
Which is why I really admire the statement in your paper trying to get compatibility with all (and seeing your efforts, you should get at least most) Nanonis STMs.
