### 2024-04-27 #6028 @ceds92
Hi @KoenImdea,

Sorry to be a pain, but since I don't have access to version 5e could you please let me know what the output is when you run the attached script with only the Z-channel highlighted?
![image](https://github.com/openjournals/joss-reviews/assets/85980355/f4558b8d-c74c-44f3-8ce3-881d16eec29f)

On version 5, the output is the following:

```

Slot number: 14
RT Idx 30
Buffer channels [14]
```

I just want to confirm the behaviour before I release the fix we discussed.

[z-channel_check.zip](https://github.com/openjournals/joss-reviews/files/15136843/z-channel_check.zip)

Thanks
