### 2024-04-29 #6028 @KoenImdea
Dear @ceds92,
I rather think it is me who is going to be a pain once you see the output of the script on my PC:

`Traceback (most recent call last):
  File "D:\Koen\Python\scanbot-4\helper_script\z-channel_check.py", line 20, in <module>
    inSlots = signalsModule.InSlotsGet()
              ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "D:\Koen\Python36\Lib\site-packages\nanonisTCP\Signals.py", line 102, in InSlotsGet
    size = self.NanonisTCP.hex_to_int32(response[idx:idx+4])
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "D:\Koen\Python36\Lib\site-packages\nanonisTCP\nanonisTCP.py", line 57, in hex_to_int32
    return struct.unpack("<i",struct.pack("I",int("0x"+h32.hex(),16)))[0]
                                              ^^^^^^^^^^^^^^^^^^^^^^
ValueError: invalid literal for int() with base 16: '0x'`

![test_survey](https://github.com/openjournals/joss-reviews/assets/51089425/2e8681c0-26fc-43a9-beae-ba2759b308ae)
