### 2024-04-29 #6028 @ceds92
Ok, thanks for doing that test @KoenImdea . I have received a reply from specs and I think I know what's going on. I'll fix it at some point this week and hopefully it's ready for when you get back
