### 2024-05-07 #6028 @ceds92
Hi @KoenImdea, I have been in communicating with specs - there is no way to obtain the version of nanonis using the TCP interface. To make Scanbot compatible with the different versions of nanonis will take some work. I need to rewrite parts of the [nanonisTCP](https://github.com/New-Horizons-SPM/nanonisTCP) library to accommodate for changes to the TCP interface across different versions. It's not a huge amount of work but I'm pretty busy in the lab at the moment so it will take me some time to get to it (hopefully ~2 weeks). I'll be tracking this fix in the issue I created above.

Anyway, just thought I'd give you this update - thanks for your patience!
