### 2024-05-09 dd813b8
Now compatible with other versions of Nanonis
Use new functions and updated nanonisTCP interface to make Scanbot compatible with different versions of Nanonis.
Can set nanonis_version in the configuration
