### 2024-05-13 de070c7
Update documentation
Removed a lot of old information
