+++ docs/commands.md
@@ -7,8 +7,8 @@
 
   gtag('config', 'G-6MK4DRHXWM');
 </script>
-# Commands
-This page documents the available commands and how to use them when running Scanbot via Zulip or in a terminal.
+# Scanbot User Guide (Terminal Mode)
+This page documents the available commands and how to use them when running Scanbot in a terminal.
 For a list of available commands on the fly, run ```help```, or for help with a specific command, run ```help <command_name>```
 <br>
 ## Scanbot Configuration
