+++ docs/web-app-test.md
@@ -24,21 +24,19 @@ and extract it to scanbot/scanbot/ before the installation</strong>. This folder
 
 ## Launching Scanbot (5 min)
 
-### Nanonis V5 Simulator
+### 1. Opening the Nanonis V5 Simulator
 1. Open the Nanonis V5 Simulator
-2. In the 'Current' module, change the Gain to LN 10^9. <strong>If your version of the simulator does not have this setting you can ignore this step.</strong>
-<br>![CurrenModule](./testing/gain109.png)<br>
-3. In the Z-Controller module, turn on the Controller. Ensure the tip is 'tunnelling' at the correct setpoint.
+2. From the Z-Controller module, enter a reasonable setpoint (10 pA - 100 pA) and turn the Controller on.
 <br>
 ![Z-ControllerModule](./testing/z-controller-on.png)
-<br>
-4. From the Scan Control module in Nanonis, ensure all images are being autosaved. <strong>This is essential for proper operation of Scanbot</strong>. It's also useful to turn on the auto-paste option too.
-<br>
-![autosave](./testing/saveall.png)
+3. If the tip height indicator does not approach to the correct setpoint, it may be because the gain setting is incorrect.
+In the 'Current' module, try changing the Gain to LN 10^9.
+<br>![CurrenModule](./testing/gain109.png)<br>
 <br>
 
-### Launch Scanbot
+### 2. Launch Scanbot
 1. Open Scanbot.exe. A terminal will open and you should see the server start.
+Alternatively, <strong>if you installed Scanbot using pip or from source, run ```scanbot```<strong>
 <br>
 ![Scanbot-terminal](./testing/scanbot-terminal.png)
 <br>
@@ -54,13 +52,15 @@ You may receive an additional warning message if Scanbot is unable to communicat
 <br>
 ![tcp-ports](./testing/tcp-ports.png)
 <br>
-2. From the Scanbot landing page, navigate to the 'Configuration' page.
-3. Ensure that the comma-delimited ports in Scanbot match those configured in Nanonis.
+2. Take note of the Nanonis version you're running from Nanonis > help > info
+![check-nanonis-version](./testing/version-help.png)
+3. In Scanbot, from the landing page, navigate to the 'Configuration' page.
+3. Ensure that the comma-delimited ports in Scanbot match those configured in Nanonis and the version numbers match.
 <br>
 ![configuration-page](./testing/configuration-page.png)
 <br>
-4. When running Scanbot using the simulator, the default configuration is ok. Accept the configuration. This will save the scanbot_config.ini file and your settings will be remembered.
-Accepting the configuration will also stop the warning message from being displayed.
+4. Accept the configuration. This will save the scanbot_config.ini file and your settings will be remembered.
+Accepting the configuration will also stop the warning message from being displayed. For details about the rest of the configuration, see the [user guide](../web-app/#configuration)
 
 ## Sample Surveying (10 min)
 1. From the Scanbot landing page, navigate to 'Data Acquisition' => 'Survey'.
@@ -89,6 +89,8 @@ In Summary:
 3.	A downward scan will acquire an image of the imprint left behind.
 4.	From this image, the imprint size and circularity will be calculated.
 5.	If the imprint size and circularity do not meet the desired criteria, a more aggressive tip-shaping action will be performed and the process repeats.
+
+<strong>Note:</strong> In demo mode, Step 1 always passes.
 <br>
 
 <strong>With Scanbot:</strong>
@@ -113,16 +115,12 @@ In the figure below, the tip was at an intial height of -2 nm, then moved 3 nm i
     * <strong>The last imprint in the demo data has an imprint size of 1.55 nm<sup>2</sup> and a circularity of 0.83.</strong>
[…]

[…]
