### 2024-05-13 1874a95
Update version 4.2.2
Version 4.2.2 has also been released to pypi
