### 2024-05-13 #6028 @ceds92
Hi @KoenImdea,

I have tested the fix for Nanonis version compatibility on the latest version of Nanonis V5 13520 (as well as my older version 11391) and they both seem to be working.

It is available for you to have another look when you're ready.
