### 2024-05-14 New-Horizons-SPM/scanbot#37 @ceds92
Scanbot review
Creating this issue for problems encountered during Scanbot's review
