### 2024-05-14 #6028 @ceds92
Hi @KoenImdea,

Thanks for trying to test again. Sorry to hear you had problems installing Scanbot - I have created a [separate issue](https://github.com/New-Horizons-SPM/scanbot/issues/37) and will try to solve it there.

Just curiously, it sounds like you attempted to install Scanbot by cloning the repository and following the instructions [here](https://new-horizons-spm.github.io/scanbot/web-app/#installation).

I usually run into that 404 error if I have forgotten to run the command in step 4:

```npm run build```

before running

```pip install .```

```npm run build``` packages the front-end into a build folder which is where the server looks for the .js files. When running ```pip install .```, that build folder should be copied to the installation location.

If you did run step 4 and this is not the case, I'll have to dig deeper.

Thanks for your time!
