### 2024-05-16 #6028 @KoenImdea
Hi @ceds92,
I think there is a curse on me. Now I just get a 404 in my browser.
This is what my CLI says:
`Successfully installed scanbot-4.4.2
PS D:\Koen\Python\scanbot-4> scanbot
Initialising app...
Loading scanbot_config.ini...
WARNING: invalid line in config file:
Loading whitelist.txt...
No whitelist found... create one with add_user
 * Serving Flask app 'scanbot.server.server'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on http://127.0.0.1:5000
Press CTRL+C to quit
127.0.0.1 - - [14/May/2024 16:52:45] "GET / HTTP/1.1" 404 -`

The problem could be with my node.js. The node command and npm install command work (and I installed your App), but I always have the feeling something is not right with them on my PC. Which is one of the reasons why I tend to develop in Docker containers. However, I do not have one suitable to run your app as I maintain a separation between the ones running node and the ones running Python.
