### 2024-05-16 #6028 @KoenImdea
Hi @ceds92,

Yes, sorry. You are right, that was a problem exists between keyboard and chair situation.

Your fix did something. Unfortunately, I'm still getting some errors.

Quoting everything would make my post way to long.
I've tried to run one survey, and one bias dependent imaging.
This is the output I get on the command line:
[PowerShell_output.txt](https://github.com/openjournals/joss-reviews/files/15332677/PowerShell_output.txt)
Both routines seem to have problems with a empty or zero-size array.

Here are the two *.sxm files as saved by my Nanonis:
[ScanbotTests.zip](https://github.com/openjournals/joss-reviews/files/15332718/ScanbotTests.zip)

Cheers,
Koen
