### 2024-05-16 #6028 @ceds92
Ok, thanks @KoenImdea. Apologies, I feel like I'm wasting a lot of your time with this.

Scanbot currently is running on two STMs here at Monash university (both with Nanonis V5) so I think there must be further incompatibilities with V5e. It's a little frustrating that I can't get a licence for V5e to test before I push through fixes hoping that they will work on your machine. I will try to contact Specs again to see if they can allow me something temporarily but I'm not sure how else to proceed here.

I'll let you know how I go.

Julian
