### 2024-05-16 #6028 @ceds92
@KoenImdea, good news... I have explained the situation to specs and they have agreed to give me a temporary license for 5e so I can solve these issues on my own instead of going back and forth.

I'll keep you posted.

Thanks
