### 2024-06-17 #6028 @ceds92
Hi @zhubonan, thanks for checking in. I had my pre-submission milestone (and passed :)) on Thursday so I literally started looking into this today.

I tested Scanbot on the latest versions of V5 and V5e (R13718) and it seems to be working. I was able to download the specific version of V5e that @KoenImdea is using (R11796) and I was able to reproduce the errors he is encountering.

So, I have narrowed down the issue to a specific version/version range at least.
I am waiting on a reply from SPECS (vendor of Nanonis) to see if they can give me the release notes for the problematic versions.
I was going to wait until they replied before posting anything here, but since you asked, this is the update.

I'll post another update when I know more.
