### 2024-06-18 #6028 @ceds92
Hi @KoenImdea,

After racking my brain for a few hours trying to figure this out, I received the below response from specs:

> I have tested Scanbot on the 5e version you gave me (thanks for that!) and it seems to be working fine with Scanbot.
> I think the issue is with a specific version (or maybe version range) where FrameDataGrab is returning an empty array for the frame data.
> This behaviour is happening on R11796 (the version the reviewer happens to be testing on) but is not happening on the latest version I have (R13718).
> Please see the attached Python script I used.
>
> I was trying to look in the release notes if there was something about this but couldn't find anything. Just wondering if you could provide some insight?
>
> Thanks,
> Julian

Response:

> Hi Julian,
>there was indeed a known issue with the API that returned the scan data.
>The indexing of the signal was wrong, so the function didn’t return any error but the returned data was wrong (or empty).
>It was fixed in the R12280 software release (November 2022).

I don't see an easy and clean way around this so I might just have to put a note in the documentation which specifies versions of Nanonis are compatible/incompatible. would it be possible to test Scanbot on R12280 or later?
