### 2024-06-26 4e5b7e5
Fix version compatibility with STS grid
STS grid no longer relies on TCP interface to get channels before STS is acquired. Instead, the channels are added as they go
