### 2024-06-26 b63db5a
Update STS grid code
Don't flipud the data
