### 2024-06-26 2b13839
Update help for STS grid - drift correction field
