### 2024-06-27 2e245fa
New numpy 2.0.0 breaks everything
keep numpy <2 until other packages become compatible with numpy 2
