+++ pyproject.toml
@@ -18,7 +18,7 @@ include = [
 
 [tool.poetry.dependencies]
 python = "^3.9"
-numpy = "*"
+numpy = "<2"
 matplotlib = "*"
 scipy = "*"
 "opencv-python" = "<4.9.0.81"
