### 2024-06-28 3599148
Automatically check for updates and minor enhancements
- web app checks an api to see if there is a new version of Scanbot available. if there is, notify the user on the home screen
- remove unused whitelist so warnings about it stop showing up
