### 2024-07-03 4ffe081
Enhance STS grid
when ramping bias, the Z-Controller is held by default. Turn this off when ramping bias during STS grids to prevent large currents.
