### 2024-07-05 #6028 @zhubonan
Hi @ceds92 there are a few more things for you to do before I can move the submission to the next stage. Please make a release of the package and deposit to a publicly available repository (for example, Zenodo) to get an DOI. This DOI will be linked to the JOSS publication.
Please see the author checklist above and make sure the metadata matches with that of the paper.

In the manuscript test, please make sure that the figures are cross-referenced (line 26 and 29 in the Draft).

PS: Other recently published JOSS papers can act as examples on how the archive DOI is linked with the paper.
