### 2024-07-06 5366d0f
Update paper.md
cross-ref figs
