### 2024-07-06 #6028 @ceds92
Hi @zhubonan,

The correct version is 4.5.1 with the following DOI: 10.5281/zenodo.12669343

I have double checked and the meta data on zenodo now matches the JOSS paper (authors, ORCIDs, and licence).

I also updated the paper so the figures are cross-referenced.

Please let me know if there's anything else
