### 2024-07-15 3239f27
Fixes to JOSS paper
