### 2024-07-15 #6028 @kyleniemeyer
Hi @ceds92, I made a few small fixes in https://github.com/New-Horizons-SPM/scanbot/pull/38, can you merge this?
